### 2024-09-16 566853d
Fix affiliation: spell out USA (#333)
