### 2024-09-16 2af864a
add JOSS badge to readme
