### 2024-09-16 #7228 @mscheltienne
@Kevin-Mattheus-Moerman Thank you for this initial checklist.

- DOIs: It looks correct to me, the `SKIP DOIs` entries don't have an associated DOIs that I could use.
- Contributing guidelines: I added the guidelines here https://github.com/mne-tools/mne-lsl/pull/332
  JOSS list the package version at `1.5.0`, and I don't see an `editorialbot` command to edit the version. Can I cut a new `1.6.0` release including the contributing guidelines or would it cause an issue for the JOSS review?
- Spelling of USA: https://github.com/mne-tools/mne-lsl/pull/333
- Particular hardware/devices to review: Nothing extra is needed, all the tutorials and examples can be run as is and `mne-lsl` includes a mock object to test and develop projects locally without specific hardware. However, if a reviewers has access to an EEG (or other) device and to its LSL application, he is more than welcome to review with this device.

Looking at the potential reviewers, I can suggested:
- adswa
- AJQuinn
- anshu-97
- vahid-sb
