### 2024-09-17 #7228 @Kevin-Mattheus-Moerman
@mscheltienne thanks for those changes. The handling editor can update the version tag later. We ask that you mint a new release at the end of the review and we'll use the version tag for that then. So there is no need to increase the version tag from our end. But if you want to increment the version that is fine and the editor can update it at any point.
