### 2024-09-17 218909f
[MNT] Release 1.6.0 (#335)
