+++ pyproject.toml
@@ -179,6 +179,7 @@ include = ['mne_lsl*']
 [tool.tomlsort]
 all = true
 ignore_case = true
+spaces_before_inline_comment = 2
 trailing_comma_inline_array = true
 
 [tool.uv.pip]
