### 2024-10-05 2d16d99
Update toml-sort config (#341)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
