+++ 1.3.0/_sources/sg_execution_times.rst.txt
@@ -1,0 +1,43 @@
+
+:orphan:
+
+.. _sphx_glr_sg_execution_times:
+
+
+Computation times
+=================
+**00:12.123** total execution time for 3 files **from all galleries**:
+
+.. container::
+
+  .. raw:: html
+
+    <style scoped>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
+    </style>
+    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
+    <script type="text/javascript" class="init">
+    $(document).ready( function () {
+        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
+    } );
+    </script>
+
+  .. list-table::
+   :header-rows: 1
+   :class: table table-striped sg-datatable
+
+   * - Example
+     - Time
+     - Mem (MB)
+   * - :ref:`sphx_glr_generated_tutorials_00_introduction.py` (``../tutorials/00_introduction.py``)
+     - 00:08.763
+     - 129.0
+   * - :ref:`sphx_glr_generated_tutorials_10_low_level_API.py` (``../tutorials/10_low_level_API.py``)
+     - 00:03.361
+     - 21.9
+   * - :ref:`sphx_glr_generated_tutorials_20_player_annotations.py` (``../tutorials/20_player_annotations.py``)
+     - 00:00.000
+     - 8.8
