+++ 1.3.0/_sources/generated/api/mne_lsl.lsl.resolve_streams.rst.txt
@@ -1,0 +1,9 @@
+﻿mne\_lsl.lsl.resolve\_streams
+=============================
+
+.. currentmodule:: mne_lsl.lsl
+
+.. autofunction:: resolve_streams
+
+.. minigallery:: mne_lsl.lsl.resolve_streams
+    :add-heading:
