### 2024-10-08 3fdd973
[pre-commit.ci] pre-commit autoupdate (#342)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
