+++ .github/workflows/pytest.yaml
@@ -15,7 +15,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu, macos, windows]
-        python-version: [3.9, "3.10", "3.11", "3.12"]
+        python-version: ["3.10", "3.11", "3.12", "3.13"]
     name: ${{ matrix.os }} - py${{ matrix.python-version }}
     runs-on: ${{ matrix.os }}-latest
     defaults:
@@ -110,7 +110,7 @@ jobs:
       fail-fast: false
       matrix:
         mne-version: ["1.4.2", "1.5.0"]
-        python-version: [3.9]
+        python-version: [3.10]
     name: mne compat ${{ matrix.mne-version }} - py${{ matrix.python-version }}
     runs-on: ubuntu-latest
     defaults:
