### 2024-10-12 01c1d3f
update doc and more ruff
