+++ mne_lsl/stream/tests/test_stream_lsl.py
@@ -67,7 +67,7 @@ def _player_mock_lsl_stream(
     player.stop()
 
 
-@pytest.fixture()
+@pytest.fixture
 def mock_lsl_stream(fname, request, chunk_size):
     """Create a mock LSL stream for testing."""
     manager = mp.Manager()
@@ -104,7 +104,7 @@ def _sleep_until_new_data(acq_delay, player):
     time.sleep(factor * max(acq_delay, player.chunk_size / player.info["sfreq"]))
 
 
-@pytest.mark.slow()
+@pytest.mark.slow
 def test_stream(mock_lsl_stream, acquisition_delay, raw):
     """Test a valid Stream."""
     # test connect/disconnect
@@ -152,7 +152,7 @@ def test_stream(mock_lsl_stream, acquisition_delay, raw):
     stream.disconnect()
 
 
-@pytest.mark.slow()
+@pytest.mark.slow
 def test_stream_invalid():
     """Test creation and connection to an invalid stream."""
     stream = Stream(bufsize=2, name="101")
@@ -210,7 +210,7 @@ def test_stream_double_connection(mock_lsl_stream):
     stream.disconnect()
 
 
-@pytest.mark.slow()
+@pytest.mark.slow
 def test_stream_drop_channels(mock_lsl_stream, acquisition_delay, raw):
     """Test dropping channels."""
     stream = Stream(
@@ -256,7 +256,7 @@ def test_stream_drop_channels(mock_lsl_stream, acquisition_delay, raw):
     stream.disconnect()
 
 
-@pytest.mark.slow()
+@pytest.mark.slow
 def test_stream_pick(mock_lsl_stream, acquisition_delay, raw):
     """Test channel selection."""
     stream = Stream(
@@ -422,7 +422,7 @@ def test_stream_channel_units(mock_lsl_stream, raw):
     stream.disconnect()
 
 
-@pytest.mark.slow()
+@pytest.mark.slow
 def test_stream_add_reference_channels(mock_lsl_stream, acquisition_delay, raw):
     """Test add reference channels and channel selection."""
     stream = Stream(
@@ -489,7 +489,7 @@ def test_stream_repr(mock_lsl_stream):
     assert stream.__repr__() == "<Stream: OFF | (source: test)>"
 
 
-@pytest.mark.slow()
+@pytest.mark.slow
 def test_stream_get_data_picks(mock_lsl_stream, acquisition_delay, raw):
     """Test channel sub-selection when getting data."""
     stream = Stream(
@@ -515,7 +515,7 @@ def test_stream_get_data_picks(mock_lsl_stream, acquisition_delay, raw):
     stream.disconnect()
 
 
-@pytest.mark.slow()
+@pytest.mark.slow
 def test_stream_n_new_samples(mock_lsl_stream, caplog):
     """Test the number of new samples available."""
     stream = Stream(
@@ -574,7 +574,7 @@ def _player_mock_lsl_stream_int(
     (player.chunk_size / player.info["sfreq"])
 
 
-@pytest.fixture()
+@pytest.fixture
 def mock_lsl_stream_int(request, chunk_size):
     """Create a mock LSL stream streaming the channel number continuously."""
     manager = mp.Manager()
@@ -597,7 +597,7 @@ def mock_lsl_stream_int(request, chunk_size):
     process.kill()
 
 
-@pytest.mark.slow()
+@pytest.mark.slow
 def test_stream_rereference(mock_lsl_stream_int, acquisition_delay):
     """Test re-referencing an EEG-like stream."""
     stream = Stream(
@@ -770,7 +770,7 @@ def _player_mock_lsl_stream_annotations(
[…]
