### 2024-10-12 e301247
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
