+++ examples/10_peak_detection.py
@@ -80,6 +80,7 @@ fig, ax = plt.subplots(1, 1, figsize=(10, 5), layout="constrained")
 for raw_, label in zip(
     (raw, raw_notched, raw_bandpassed, raw_lowpassed),
     ("raw", "notched", "bandpassed", "lowpassed"),
+    strict=False,
 ):
     data, times = raw_[:, start:stop]  # select 5 seconds
     data -= data.mean()  # detrend
@@ -97,6 +98,7 @@ fig, ax = plt.subplots(1, 1, figsize=(10, 5), layout="constrained")
 for raw_, label in zip(
     (raw, raw_notched, raw_bandpassed, raw_lowpassed),
     ("raw", "notched", "bandpassed", "lowpassed"),
+    strict=False,
 ):
     data, times = raw_[:, start:stop]  # select 5 seconds
     data -= data.mean()  # detrend
