+++ mne_lsl/stream/base.py
@@ -112,7 +112,7 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
     def add_reference_channels(
         self,
         ref_channels: str | list[str] | tuple[str],
-        ref_units: str | int | list[str | int] | tuple[str | int] | None = None,
+        ref_units: str | int | list[str | int] | tuple[str | int, ...] | None = None,
     ) -> BaseStream:
         """Add EEG reference channels to data that consists of all zeros.
 
