+++ .github/workflows/pytest.yaml
@@ -110,7 +110,7 @@ jobs:
       fail-fast: false
       matrix:
         mne-version: ["1.4.2", "1.5.0"]
-        python-version: [3.10]
+        python-version: ["3.10"]
     name: mne compat ${{ matrix.mne-version }} - py${{ matrix.python-version }}
     runs-on: ubuntu-latest
     defaults:
