### 2024-10-16 59a221f
[MNT] Drop python 3.9 and support python 3.13 (#343)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
