+++ mne_lsl/utils/tests/test_logs.py
@@ -9,7 +9,6 @@ from mne_lsl.utils.logs import _use_log_level, add_file_handler, logger, verbose
 
 if TYPE_CHECKING:
     from pathlib import Path
-    from typing import Optional, Union
 
 
 def test_default_log_level(caplog: pytest.LogCaptureFixture):
@@ -68,7 +67,7 @@ def test_verbose(caplog: pytest.LogCaptureFixture):
 
     # function
     @verbose
-    def foo(verbose: Optional[Union[bool, str, int]] = None):
+    def foo(verbose: bool | str | int | None = None):
         """Foo function."""
         logger.debug("101")
 
@@ -94,12 +93,12 @@ def test_verbose(caplog: pytest.LogCaptureFixture):
             pass
 
         @verbose
-        def foo(self, verbose: Optional[Union[bool, str, int]] = None):
+        def foo(self, verbose: bool | str | int | None = None):
             logger.debug("101")
 
         @staticmethod
         @verbose
-        def foo2(verbose: Optional[Union[bool, str, int]] = None):
+        def foo2(verbose: bool | str | int | None = None):
             logger.debug("101")
 
     foo = Foo()
