### 2024-10-21 1d48e5a
safer parse of channel names
