+++ doc/development/changes/1.6.rst
@@ -22,3 +22,4 @@ Version 1.6
 - Improve total code coverage (:pr:`317` by `Mathieu Scheltienne`_)
 - Add mailmap to the repository (:pr:`327` by `Daniel McCloy`_)
 - Fix indexing when replaying datasets from non-preloaded :class:`~mne.io.Raw` (:pr:`330` by `Toni M. Brotons`_)
+- Safer parse of channel names in the legacy stream viewer (by `Mathieu Scheltienne`_)
