+++ pyproject.toml
@@ -18,6 +18,7 @@ classifiers = [
   'Programming Language :: Python :: 3.10',
   'Programming Language :: Python :: 3.11',
   'Programming Language :: Python :: 3.12',
+  'Programming Language :: Python :: 3.13',
   'Programming Language :: Python :: 3.9',
   'Topic :: Scientific/Engineering',
   'Topic :: Software Development',
@@ -55,7 +56,7 @@ maintainers = [
 name = 'mne_lsl'
 readme = 'README.md'
 requires-python = '>=3.9'
-version = '1.6.0'
+version = '1.6.1'
 
 [project.optional-dependencies]
 all = [
