### 2024-10-25 6ec1b6e
hotfix for 64/128 channel ANT amplifiers with CB eego2lsl
