+++ .github/workflows/doc.yaml
@@ -26,10 +26,10 @@ jobs:
         uses: ./.github/actions/install-system-dependencies
       - name: Install optipng (for sphinx-gallery)
         run: sudo apt install optipng
-      - name: Install uv and package
-        run: |
-          python -m pip install --quiet uv
-          uv pip install --quiet --system .[doc]
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
+      - name: Install package
+        run: uv pip install --quiet --system .[doc]
       - name: Display system information
         run: mne-lsl sys-info --developer
       - name: Get dataset
