+++ .github/workflows/publish.yaml
@@ -17,10 +17,10 @@ jobs:
           python-version: 3.11
       - name: Install system dependencies
         uses: ./.github/actions/install-system-dependencies
-      - name: Install uv and package
-        run: |
-          python -m pip install --quiet uv
-          uv pip install --quiet --system -e .[build,stubs]
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
+      - name: Install package
+        run: uv pip install --quiet --system -e .[build,stubs]
       - name: Display system information
         run: mne-lsl sys-info --developer
       - name: Generate stub files
