### 2024-10-31 b090f4d
improve type hints
