### 2024-10-31 1e5a57d
[MNT] Use more uv in workflows (#349)
