+++ .github/workflows/pytest.yaml
@@ -30,10 +30,10 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install system dependencies
         uses: ./.github/actions/install-system-dependencies
-      - name: Install uv and package
-        run: |
-          python -m pip install --quiet uv
-          uv pip install --quiet --system .[test]
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
+      - name: Install package
+        run: uv pip install --quiet --system .[test]
       - name: Display system information
         run: mne-lsl sys-info --developer
       - name: Get testing dataset
@@ -75,9 +75,10 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install system dependencies
         uses: ./.github/actions/install-system-dependencies
-      - name: Install uv and package
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
+      - name: Install package
         run: |
-          python -m pip install --quiet uv
           uv pip install --quiet --system .[test]
           uv pip install --quiet --system --upgrade git+https://github.com/mne-tools/mne-python
           uv pip install --quiet --system matplotlib
@@ -125,11 +126,12 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install system dependencies
         uses: ./.github/actions/install-system-dependencies
-      - name: Install uv and package
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
+      - name: Install package
         run: |
-          python -m pip install --quiet uv
           uv pip install --quiet --system .[test]
-          python -m pip uninstall numpy mne -y
+          uv pip uninstall --quiet --system numpy mne
           uv pip install --quiet --system "numpy<2"
           uv pip install --quiet --system mne==${{ matrix.mne-version }}
       - name: Display system information
