### 2024-10-31 20a9bf6
add config variable to set the libfolder download target
