### 2024-10-31 2a493fc
Add configuration variable MNE_LSL_LIB_FOLDER to control the download directory (#350)
