+++ .github/workflows/build_wheels.yml
@@ -11,10 +11,13 @@ jobs:
     name: Build wheels on ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
     strategy:
+      fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest, macos-13, macos-14]
     steps:
       - uses: actions/checkout@v4
+        with:
+          submodules: recursive
       - uses: pypa/cibuildwheel@v2.21.3
       - uses: actions/upload-artifact@v4
         with:
