+++ tools/cibw_before_all.sh
@@ -11,8 +11,13 @@ if [[ "$RUNNER_OS" == 'Linux' ]]; then
     echo "Compiling on Linux"
 elif [[ "$RUNNER_OS" == 'macOS' ]]; then
     echo "Compiling on macOS"
+    # On arm64, we need to set the install path to /opt/local/lib
+    if [[ "$(uname -m)" == "arm64" ]]; then
+        export CMAKE_INSTALL_PREFIX=/opt/local
+    fi
 elif [[ "$RUNNER_OS" == 'Windows' ]]; then
     echo "Compiling on Windows"
+    export CMAKE_INSTALL_PREFIX="C:/Program Files/liblsl"
 else
     echo "Unknown RUNNER_OS: ${RUNNER_OS}"
     exit 1
