+++ .github/workflows/build_wheels.yml
@@ -18,6 +18,10 @@ jobs:
       - uses: actions/checkout@v4
         with:
           submodules: recursive
+      - run: |
+          sudo mkdir -p /usr/local/lib /usr/local/include
+          sudo chmod a+rwx /usr/local/lib /usr/local/include
+        if: matrix.os == 'macos-14'
       - uses: pypa/cibuildwheel@v2.21.3
       - uses: actions/upload-artifact@v4
         with:
