+++ tools/cibw_before_all.sh
@@ -11,7 +11,6 @@ if [[ "$RUNNER_OS" == 'Linux' ]]; then
     echo "Compiling on Linux"
 elif [[ "$RUNNER_OS" == 'macOS' ]]; then
     echo "Compiling on macOS"
-    # On arm64, we need to set the install path to /opt/local/lib
 elif [[ "$RUNNER_OS" == 'Windows' ]]; then
     echo "Compiling on Windows"
     export CMAKE_INSTALL_PREFIX="C:/Program Files/liblsl"
