+++ pyproject.toml
@@ -138,8 +138,7 @@ inherit.repair-wheel-command = "append"
 repair-wheel-command = "delvewheel repair -w {dest_dir} {wheel}"
 
 [tool.cibuildwheel.windows.environment]
-INCLUDE = "$GITHUB_WORKSPACE\\liblsl\\include"
-LIB = "$GITHUB_WORKSPACE\\liblsl\\build"
+CL = "-I%GITHUB_WORKSPACE%\\liblsl\\include -L%GITHUB_WORKSPACE%\\liblsl\\Release"
 
 [tool.codespell]
 check-filenames = true
