+++ tools/cibw_before_all.sh
@@ -13,8 +13,6 @@ elif [[ "$RUNNER_OS" == 'macOS' ]]; then
     echo "Compiling on macOS"
 elif [[ "$RUNNER_OS" == 'Windows' ]]; then
     echo "Compiling on Windows"
-    export CMAKE_INSTALL_PREFIX="C:/Program Files/liblsl"
-    TARGET_OPTION="--target install"
 else
     echo "Unknown RUNNER_OS: ${RUNNER_OS}"
     exit 1
@@ -24,11 +22,11 @@ cd liblsl
 # https://labstreaminglayer.readthedocs.io/dev/lib_dev.html#configuring-the-liblsl-project
 set -x
 cmake -S . -B build -DCMAKE_BUILD_TYPE=Release
-cmake --build build --config release $TARGET_OPTION
+cmake --build build --config release
 echo "Done building liblsl:"
 ls -alt build
-set +x
 if [[ "$RUNNER_OS" == 'Windows' ]]; then
-    ls -alR C:/Program\ Files/liblsl
+    ls -alR build/Release
 fi
+set +x
 cd ..
