+++ .github/workflows/build_wheels.yml
@@ -18,11 +18,14 @@ jobs:
       - uses: actions/checkout@v4
         with:
           submodules: recursive
-      - run: |
-          sudo mkdir -p /usr/local
-          sudo chown $USER /usr/local
-          mkdir -p /usr/local/lib /usr/local/include
-        if: matrix.os == 'macos-14'
+      # For some reason doing this in pyproject.toml does not seem to work :(
+      - name: Set Windows compilation env vars
+        shell: bash
+        run: |
+          set -exo pipefail
+          echo "INCLUDE=$GITHUB_WORKSPACE\\liblsl\\include" >> $GITHUB_ENV
+          echo "LIB=$GITHUB_WORKSPACE\\liblsl\\build\\Release" >> $GITHUB_ENV
+        if: runner.os == 'Windows'
       - uses: pypa/cibuildwheel@v2.21.3
       - uses: actions/upload-artifact@v4
         with:
