+++ .github/workflows/build_wheels.yml
@@ -1,4 +1,7 @@
-name: Build
+name: Wheels
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.number }}-${{ github.event.ref }}
+  cancel-in-progress: true
 
 on:  # yamllint disable-line rule:truthy
   pull_request:
