+++ pyproject.toml
@@ -122,7 +122,7 @@ RUNNER_OS = "Linux"  # could use environment-pass but this facilitates local tes
 [tool.cibuildwheel.macos]
 repair-wheel-command = """\
 DYLD_LIBRARY_PATH=$GITHUB_WORKSPACE/liblsl/build \
-delocate-wheel --require-archs {delocate_archs} -w {dest_dir} -v {wheel} \
+delocate-wheel -w {dest_dir} -v {wheel} --require-archs {delocate_archs} \
 """
 
 [tool.cibuildwheel.macos.environment]
@@ -135,7 +135,7 @@ select = "cp*"  # abi3 not used for PyPi / pp*
 inherit.repair-wheel-command = "append"
 
 [tool.cibuildwheel.windows]
-repair-wheel-command = "delvewheel repair --add-path=\"$ADD_PATH\" -w {dest_dir} {wheel}"
+repair-wheel-command = "delvewheel repair --add-path=\"%ADD_PATH%\" -w {dest_dir} {wheel}"
 
 [tool.codespell]
 check-filenames = true
