+++ .github/workflows/build_wheels.yml
@@ -28,7 +28,7 @@ jobs:
           set -exo pipefail
           echo "INCLUDE=$GITHUB_WORKSPACE\\liblsl\\include" | tee -a $GITHUB_ENV
           echo "LIB=$GITHUB_WORKSPACE\\liblsl\\build\\Release" | tee -a $GITHUB_ENV
-          echo "ADD_PATH=$GITHUB_WORKSPACE\\liblsl\\build\\Release" | tee -a $GITHUB_ENV
+          echo "ADD_PATH=$(cygpath -w $GITHUB_WORKSPACE)\\liblsl\\build\\Release" | tee -a $GITHUB_ENV
         if: runner.os == 'Windows'
       - uses: pypa/cibuildwheel@v2.21.3
       - uses: actions/upload-artifact@v4
