+++ .github/workflows/build_wheels.yml
@@ -48,6 +48,7 @@ jobs:
       - uses: actions/download-artifact@v4
         with:
           pattern: cibw-wheels-${{ matrix.os }}-*
+          merge-multiple: true
       - uses: actions/setup-python@v5
         with:
           python-version: 3.12
