+++ .github/workflows/build_wheels.yml
@@ -49,6 +49,7 @@ jobs:
         with:
           pattern: cibw-wheels-${{ matrix.os }}-*
           merge-multiple: true
+          path: wheelhouse
       - uses: actions/setup-python@v5
         with:
           python-version: 3.12
@@ -56,6 +57,7 @@ jobs:
         shell: bash
         run: |
           set -eo pipefail
+          ls -alR wheelhouse
           FNAME=$PWD/wheelhouse/*.whl
           echo "FNAME=$FNAME"
           test -f $FNAME
