+++ .github/workflows/build_wheels.yml
@@ -74,7 +74,7 @@ jobs:
       - name: Run pytest
         uses: ./.github/actions/retry-step
         with:
-          command: pytest mne_lsl --cov=mne_lsl --cov-report=xml --cov-config=pyproject.toml -s "$TEST_PATH"
+          command: pytest "$TEST_PATH" --cov=mne_lsl --cov-report=xml --cov-config=pyproject.toml -s
           retry_error_codes: "3,134,139"
         env:
           MNE_LSL_LOG_LEVEL: DEBUG
