### 2024-11-08 fe21a16
set RUNNER_OS everywhere
