### 2024-11-08 550c95e
simplify load_liblsl
