+++ mne_lsl/lsl/tests/test_load_liblsl.py
@@ -1,174 +1,28 @@
-import platform
-from ctypes import c_void_p, sizeof
-from pathlib import Path
-from shutil import copy
-
-import pooch
 import pytest
 
 from mne_lsl.lsl.load_liblsl import (
-    _PLATFORM,
-    _PLATFORM_SUFFIXES,
-    _attempt_load_liblsl,
-    _fetch_liblsl,
-    _is_valid_libpath,
-    _is_valid_version,
-    _load_liblsl_mne_lsl,
-    _pooch_processor_liblsl,
+    _load_liblsl_environment_variables,
+    _load_liblsl_wheel_path,
 )
-
-
-@pytest.fixture(scope="module")
-def download_liblsl_outdated(tmp_path_factory) -> Path:
-    """Fixture to download an outdated liblsl version."""
-    assert _PLATFORM in _PLATFORM_SUFFIXES  # test OS-detection
-    if _PLATFORM == "darwin" and platform.processor() == "i386":
-        asset = dict(
-            name="liblsl-1.14.0-OSX_amd64.tar.bz2",
-            browser_download_url="https://github.com/sccn/liblsl/releases/download/v1.14.0/liblsl-1.14.0-OSX_amd64.tar.bz2",  # noqa: E501
-            known_hash="c1f9004243db49885b18884b39d793f66c94e45d52a7bde12c51893e74db337d",  # noqa: E501
-        )
-    elif _PLATFORM == "windows" and sizeof(c_void_p) == 8:
-        asset = dict(
-            name="liblsl-1.14.0-Win_amd64.zip",
-            browser_download_url="https://github.com/sccn/liblsl/releases/download/v1.14.0/liblsl-1.14.0-Win_amd64.zip",  # noqa: E501
-            known_hash="75ec445e9e9b23b15400322fffa06666098fce42e706afa763e77abdbea87e52",  # noqa: E501
-        )
-    elif _PLATFORM == "linux":
-        import distro
-
-        if distro.codename() == "bionic":
-            asset = dict(
-                name="liblsl-1.14.0-bionic_amd64.deb",
-                browser_download_url="https://github.com/sccn/liblsl/releases/download/v1.14.0/liblsl-1.14.0-bionic_amd64.deb",  # noqa: E501
-                known_hash="b2b882414a73acba4e7ab14361a6d541cc1a3774536e08471b346a89b4b557bb",  # noqa: E501
-            )
-        elif distro.codename() == "focal":
-            asset = dict(
-                name="liblsl-1.14.0-focal_amd64.deb",
-                browser_download_url="https://github.com/sccn/liblsl/releases/download/v1.14.0/liblsl-1.14.0-focal_amd64.deb",  # noqa: E501
-                known_hash="b3a0c0e40746f126d77212400e1e23b2f3225bf0458215731b08b1bde631f15b",  # noqa: E501
-            )
-        else:
-            pytest.skip(reason=f"Unsupported linux distribution '{distro.codename()}'.")
-    else:
-        pytest.skip(reason=f"Unsupported platform '{_PLATFORM}'.")
-
-    try:
-        libpath = pooch.retrieve(
-            url=asset["browser_download_url"],
-            fname=asset["name"],
-            path=tmp_path_factory.mktemp("liblsl_outdated"),
-            processor=_pooch_processor_liblsl,
-            known_hash=asset["known_hash"],
-        )
-    except ValueError:
-        pytest.skip(reason="Unable to download the outdated liblsl (invalid hash).")
-    return Path(libpath)
 
 
 @pytest.fixture
-def liblsl_outdated(tmp_path, download_liblsl_outdated) -> Path:
-    """Fixture to provide an outdated liblsl version."""
-    copy(download_liblsl_outdated, tmp_path / download_liblsl_outdated.name)
[…]
