+++ mne_lsl/lsl/load_liblsl.py
@@ -1,47 +1,12 @@
 from __future__ import annotations
 
-import glob
 import os
 import platform
-import subprocess
-import tarfile
-import zipfile
-from ctypes import CDLL, c_char_p, c_double, c_long, c_void_p, sizeof
-from ctypes.util import find_library
-from pathlib import Path
-from shutil import move, rmtree
-from typing import TYPE_CHECKING
+from ctypes import CDLL, c_char_p, c_double, c_long, c_void_p
+from importlib.resources import files
 
-import pooch
-import requests
-from mne import get_config, set_config
-
-from .._version import __version__
 from ..utils._checks import ensure_path
-from ..utils._path import walk
 from ..utils.logs import logger, warn
-
-if TYPE_CHECKING:
-    from pooch import Pooch
-
-
-# minimum liblsl version. The major version is given by version // 100
-# and the minor version is given by version % 100.
-_VERSION_MIN: int = 115
-# liblsl objects created with the same protocol version are inter-compatible.
-_VERSION_PROTOCOL: int = 110  # noqa: W0612
-_PLATFORM: str = platform.system().lower().strip()
-_PLATFORM_SUFFIXES: dict[str, str] = {
-    "windows": ".dll",
-    "darwin": ".dylib",
-    "linux": ".so",
-}
-# generic error message
-_ERROR_MSG: str = (
-    "Please visit LIBLSL library github page (https://github.com/sccn/liblsl) and "
-    "install a release in the system directories or provide its path in the "
-    "environment variable MNE_LSL_LIB or PYLSL_LIB."
-)
 
 
 def load_liblsl() -> CDLL:
@@ -54,49 +19,21 @@ def load_liblsl() -> CDLL:
     3. Search in the defined library folder.
     4. Fetch on GitHub.
     """
-    if _PLATFORM not in _PLATFORM_SUFFIXES:  # pragma: no cover
-        raise RuntimeError(
-            "The OS could not be determined. Please open an issue on GitHub and "
-            "provide the error traceback to the developers."
-        )
     libpath = _load_liblsl_environment_variables()
-    libpath = _load_liblsl_wheel_path() if libpath is None else libpath
-    libpath = _load_liblsl_system() if libpath is None else libpath
-    libpath = _load_liblsl_mne_lsl() if libpath is None else libpath
-    libpath = _fetch_liblsl() if libpath is None else libpath
+    libpath = _load_liblsl_wheel_path()
     assert isinstance(libpath, str)  # sanity-check
     lib = CDLL(libpath)
     _set_types(lib)
     return lib
 
 
-def _load_liblsl_wheel_path() -> str | None:
-    """Load the binary LSL library from the wheel path."""
-    loc: str | None = None
-    if platform.system() == "Linux":
-        # auditwheel will relocate and mangle, e.g.:
-        # mne_lsl/../mne_lsl.libs/liblsl-65106c22.so.1.16.2
-        libs = Path(__file__).parents[2] / "mne_lsl.libs"
-        if libs.is_dir():
-            loc = glob.glob(str(libs / "liblsl*.so*"))[0]
-    elif platform.system() == "Windows":
-        # delvewheel has similar behavior to auditwheel
-        libs = Path(__file__).parents[2] / "mne_lsl.libs"
-        if libs.is_dir():
-            loc = glob.glob(str(libs / "lsl*.dll"))[0]
-    elif platform.system() == "Darwin":
-        libs = Path(__file__).parents[1] / ".dylibs"
-        if libs.is_dir():
-            loc = glob.glob(str(libs / "liblsl*.dylib"))[0]
-    if loc is not None:
-        logger.debug(f"Found wheel path {loc}")
-    else:
[…]
