+++ .github/workflows/pytest.yaml
@@ -1,156 +1,0 @@
-name: pytest
-concurrency:
-  group: ${{ github.workflow }}-${{ github.event.number }}-${{ github.event.ref }}
-  cancel-in-progress: true
-on:  # yamllint disable-line rule:truthy
-  pull_request:
-  push:
-    branches: [main]
-  workflow_dispatch:
-
-jobs:
-  pytest:
-    timeout-minutes: 30
-    strategy:
-      fail-fast: false
-      matrix:
-        os: [ubuntu, macos, windows]
-        python-version: ["3.10", "3.11", "3.12", "3.13"]
-    name: ${{ matrix.os }} - py${{ matrix.python-version }}
-    runs-on: ${{ matrix.os }}-latest
-    defaults:
-      run:
-        shell: bash
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-      - name: Setup Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Install system dependencies
-        uses: ./.github/actions/install-system-dependencies
-      - name: Install uv
-        uses: astral-sh/setup-uv@v3
-      - name: Install package
-        run: uv pip install --quiet --system .[test]
-      - name: Display system information
-        run: mne-lsl sys-info --developer
-      - name: Get testing dataset
-        uses: ./.github/actions/get-testing-dataset
-      - name: Run pytest
-        uses: ./.github/actions/retry-step
-        with:
-          command: pytest mne_lsl --cov=mne_lsl --cov-report=xml --cov-config=pyproject.toml -s
-          retry_error_codes: "3,134,139"
-        env:
-          MNE_LSL_LOG_LEVEL: DEBUG
-      - name: Upload to codecov
-        uses: codecov/codecov-action@v4
-        with:
-          files: ./coverage.xml
-          flags: unittests  # optional
-          name: codecov-umbrella  # optional
-          fail_ci_if_error: false  # optional (default = false)
-          token: ${{ secrets.CODECOV_TOKEN }}
-          verbose: true  # optional (default = false)
-
-  pytest-pip-pre:
-    timeout-minutes: 30
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.11"]
-    name: pip pre-release - py${{ matrix.python-version }}
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        shell: bash
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-      - name: Setup Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Install system dependencies
-        uses: ./.github/actions/install-system-dependencies
-      - name: Install uv
-        uses: astral-sh/setup-uv@v3
-      - name: Install package
-        run: |
-          uv pip install --quiet --system .[test]
-          uv pip install --quiet --system --upgrade git+https://github.com/mne-tools/mne-python
-          uv pip install --quiet --system matplotlib
-          uv pip install --quiet --system --upgrade --prerelease allow --only-binary :all: -i https://pypi.anaconda.org/scientific-python-nightly-wheels/simple --no-deps matplotlib
-          uv pip install --quiet --system --upgrade --prerelease allow --only-binary :all: -i https://pypi.anaconda.org/scientific-python-nightly-wheels/simple numpy scipy
-      - name: Display system information
-        run: mne-lsl sys-info --developer
-      - name: Get testing dataset
-        uses: ./.github/actions/get-testing-dataset
-      - name: Run pytest
-        uses: ./.github/actions/retry-step
[…]
