### 2024-11-08 637c4ed
fix paths and improve RuntimeError
