+++ mne_lsl/lsl/load_liblsl.py
@@ -79,23 +79,29 @@ def _load_liblsl_wheel_path() -> str:
     if platform.system() == "Linux":
         # auditwheel will relocate and mangle, e.g.:
         # mne_lsl/../mne_lsl.libs/liblsl-65106c22.so.1.16.2
-        libs = files("mne_lsl").parent / "libs"
+        libs = files("mne_lsl").parent / "mne_lsl.libs"
         lib_files = list(libs.glob("liblsl*.so*"))
         if len(lib_files) != 1:
-            raise RuntimeError("Could not find the LIBLSL library bundle with mne-lsl.")
+            raise RuntimeError(
+                f"Could not find the LIBLSL library bundle with mne-lsl in '{libs}'."
+            )
         libpath = lib_files[0]
     elif platform.system() == "Windows":
         # delvewheel has similar behavior to auditwheel
-        libs = files("mne_lsl").parent / "libs"
+        libs = files("mne_lsl").parent / "mne_lsl.libs"
         lib_files = list(libs.glob("lsl*.dll"))
         if len(lib_files) != 1:
-            raise RuntimeError("Could not find the LIBLSL library bundle with mne-lsl.")
+            raise RuntimeError(
+                f"Could not find the LIBLSL library bundle with mne-lsl in '{libs}'."
+            )
         libpath = lib_files[0]
     elif platform.system() == "Darwin":
         libs = files("mne_lsl") / ".dylibs"
         lib_files = list(libs.glob("liblsl*.dylib"))
         if len(lib_files) != 1:
-            raise RuntimeError("Could not find the LIBLSL library bundle with mne-lsl.")
+            raise RuntimeError(
+                f"Could not find the LIBLSL library bundle with mne-lsl in '{libs}'."
+            )
         libpath = lib_files[0]
     else:
         raise RuntimeError(
