+++ .github/workflows/ci.yaml
@@ -44,6 +44,12 @@ jobs:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-13, macos-latest]
         python: ["3.10", "3.13"]
+        exclude:
+          - os: macos-13
+            python: "3.13"
+    defaults:
+      run:
+        shell: bash
     steps:
       - run: sudo apt update && sudo apt install -y binutils libpugixml-dev qtbase5-dev qt5-qmake
         if: runner.os == 'Linux'
@@ -56,10 +62,9 @@ jobs:
           merge-multiple: true
           path: dist
       - run: ls -alt . && ls -alt dist/
-      - run: pip install --only-binary antio antio[test] --find-links dist --pre
+      - run: pip install --only-binary mne-lsl mne-lsl[test] --find-links dist --pre
       - run: mne-lsl sys-info --developer
       - name: Get installed library location
-        shell: bash
         run: |
           set -eo pipefail
           MODULE_PATH=$(python -c "import pathlib, mne_lsl; print(pathlib.Path(mne_lsl.__file__).parent)")
@@ -67,7 +72,6 @@ jobs:
           echo "MODULE_PATH=$MODULE_PATH" | tee -a $GITHUB_ENV
       - uses: actions/checkout@v4
       - run: rm -Rf mne_lsl
-        shell: bash
       - uses: ./.github/actions/get-testing-dataset
       - name: Run pytest
         uses: ./.github/actions/retry-step
