### 2024-11-11 bd38e16
simplify workflow
