+++ .github/workflows/ci.yaml
@@ -53,6 +53,7 @@ jobs:
     steps:
       - run: sudo apt update && sudo apt install -y binutils libpugixml-dev qtbase5-dev qt5-qmake
         if: runner.os == 'Linux'
+      - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python }}
@@ -64,19 +65,11 @@ jobs:
       - run: ls -alt . && ls -alt dist/
       - run: pip install --only-binary mne-lsl mne-lsl[test] --find-links dist --pre
       - run: mne-lsl sys-info --developer
-      - name: Get installed library location
-        run: |
-          set -eo pipefail
-          MODULE_PATH=$(python -c "import pathlib, mne_lsl; print(pathlib.Path(mne_lsl.__file__).parent)")
-          MODULE_PATH=$(echo "$MODULE_PATH" | sed 's/\\/\\\\/g')
-          echo "MODULE_PATH=$MODULE_PATH" | tee -a $GITHUB_ENV
-      - uses: actions/checkout@v4
-      - run: rm -Rf mne_lsl
       - uses: ./.github/actions/get-testing-dataset
       - name: Run pytest
         uses: ./.github/actions/retry-step
         with:
-          command: pytest "$MODULE_PATH" --cov=mne_lsl --cov-report=xml --cov-config=pyproject.toml -s
+          command: pytest tests/ --cov=mne_lsl --cov-report=xml --cov-config=pyproject.toml -s
           retry_error_codes: "3,134,139"
         env:
           MNE_LSL_LOG_LEVEL: DEBUG
