### 2024-11-11 f372bc9
fix macOS build target
