+++ .github/workflows/ci.yaml
@@ -26,9 +26,9 @@ jobs:
         shell: bash
         run: |
           set -exo pipefail
-          echo "INCLUDE=$GITHUB_WORKSPACE\\liblsl\\include" | tee -a $GITHUB_ENV
-          echo "LIB=$GITHUB_WORKSPACE\\liblsl\\build\\Release" | tee -a $GITHUB_ENV
-          echo "ADD_PATH=$(cygpath -w $GITHUB_WORKSPACE)\\liblsl\\build\\Release" | tee -a $GITHUB_ENV
+          echo "INCLUDE=$GITHUB_WORKSPACE\\src\\liblsl\\include" | tee -a $GITHUB_ENV
+          echo "LIB=$GITHUB_WORKSPACE\\src\\liblsl\\build\\Release" | tee -a $GITHUB_ENV
+          echo "ADD_PATH=$(cygpath -w $GITHUB_WORKSPACE)\\src\\liblsl\\build\\Release" | tee -a $GITHUB_ENV
         if: runner.os == 'Windows'
       - uses: pypa/cibuildwheel@v2.21.3
       - uses: actions/upload-artifact@v4
