### 2024-11-11 36a27bc
fix loading of liblsl
