### 2024-11-11 b1a695f
set minimum macOS deployment target to 11
