### 2024-11-11 fcf4122
try to use setup.py to build liblsl
