+++ pyproject.toml
@@ -1,6 +1,6 @@
 [build-system]
 build-backend = 'setuptools.build_meta'
-requires = ['cffi >= 1.0.0', 'setuptools >= 75.0.0']
+requires = ['setuptools >= 64.0.0']
 
 [project]
 authors = [
@@ -103,11 +103,8 @@ tracker = 'https://github.com/mne-tools/mne-lsl/issues'
 
 [tool.cibuildwheel]
 archs = "native"
-before-all = "bash {project}/tools/cibw_before_all.sh"
 build = ["cp310-*"]  # abi3 wheel, can build for a single Python version
 skip = ["*-musllinux*", "pp*"]  # no PyPy support for pp310 for SciPy
-test-command = "python -u {project}/tools/cibw_test_command.py"
-test-requires = ["cffi"]  # we check that the token CFFI binding works
 
 [tool.cibuildwheel.linux]
 repair-wheel-command = """\
@@ -118,8 +115,6 @@ auditwheel repair -w {dest_dir} {wheel} \
 [tool.cibuildwheel.linux.environment]
 C_INCLUDE_PATH = "/project/src/liblsl/include"  # GCC variable
 LIBRARY_PATH = "/project/src/liblsl/build"  # GCC variable required for CFFI build
-MNE_LSL_BUILD_FFI = "1"
-RUNNER_OS = "Linux"  # could use environment-pass but this facilitates local testing
 
 [tool.cibuildwheel.macos]
 repair-wheel-command = """\
@@ -131,8 +126,6 @@ delocate-wheel -w {dest_dir} -v {wheel} --require-archs {delocate_archs} \
 CPATH = "$GITHUB_WORKSPACE/src/liblsl/include"  # Clang variable
 LIBRARY_PATH = "$GITHUB_WORKSPACE/src/liblsl/build"  # Clang variable required for CFFI build
 MACOSX_DEPLOYMENT_TARGET = "11"
-MNE_LSL_BUILD_FFI = "1"
-RUNNER_OS = "macOS"
 
 [[tool.cibuildwheel.overrides]]
 repair-wheel-command = "pipx run abi3audit --strict --report {wheel}"
@@ -142,10 +135,6 @@ inherit.repair-wheel-command = "append"
 [tool.cibuildwheel.windows]
 before-build = "pip install delvewheel"
 repair-wheel-command = "delvewheel repair --add-path=\"%ADD_PATH%\" -w {dest_dir} {wheel}"
-
-[tool.cibuildwheel.windows.environment]
-MNE_LSL_BUILD_FFI = "1"
-RUNNER_OS = "Windows"
 
 [tool.codespell]
 check-filenames = true
