### 2024-11-11 c7b3c56
cast Path.glob to list
