+++ setup.py
@@ -46,7 +46,7 @@ class build_ext(_build_ext):  # noqa: D101
                 build_dir = Path(build_dir) / "Release"
             else:
                 build_dir = Path(build_dir)
-            lib_files = build_dir.glob(_PATTERN[platform.system()])
+            lib_files = list(build_dir.glob(_PATTERN[platform.system()]))
             assert len(lib_files) == 1  # sanity-check
             dst = Path(self.build_lib) / "mne_lsl" / "lsl" / "lib" / lib_files[0]
             dst.parent.mkdir(parents=True, exist_ok=True)
