+++ pyproject.toml
@@ -107,20 +107,18 @@ build = ["cp310-*"]  # abi3 wheel, can build for a single Python version
 skip = ["*-musllinux*", "pp*"]  # no PyPy support for pp310 for SciPy
 
 [tool.cibuildwheel.linux]
-repair-wheel-command = """\
-LD_LIBRARY_PATH=/project/src/liblsl/build \
-auditwheel repair -w {dest_dir} {wheel} \
-"""
+repair-wheel-command = [
+  "auditwheel repair -w {dest_dir} {wheel}",
+]
 
 [tool.cibuildwheel.linux.environment]
 C_INCLUDE_PATH = "/project/src/liblsl/include"  # GCC variable
 LIBRARY_PATH = "/project/src/liblsl/build"  # GCC variable required for CFFI build
 
 [tool.cibuildwheel.macos]
-repair-wheel-command = """\
-DYLD_LIBRARY_PATH=$GITHUB_WORKSPACE/src/liblsl/build \
-delocate-wheel -w {dest_dir} -v {wheel} --require-archs {delocate_archs} \
-"""
+repair-wheel-command = [
+  "delocate-wheel --require-archs {delocate_archs} -w {dest_dir} -v {wheel}",
+]
 
 [tool.cibuildwheel.macos.environment]
 CPATH = "$GITHUB_WORKSPACE/src/liblsl/include"  # Clang variable
@@ -134,7 +132,9 @@ inherit.repair-wheel-command = "append"
 
 [tool.cibuildwheel.windows]
 before-build = "pip install delvewheel"
-repair-wheel-command = "delvewheel repair --add-path=\"%ADD_PATH%\" -w {dest_dir} {wheel}"
+repair-wheel-command = [
+  "delvewheel repair -w {dest_dir} {wheel}",
+]
 
 [tool.codespell]
 check-filenames = true
