+++ MANIFEST.in
@@ -1,1 +1,0 @@
-exclude mne_lsl/conftest.py
