### 2024-11-11 062223d
fix path and setuptools include/exclude
