### 2024-11-11 164347e
simplify liblsl loading
