### 2024-11-12 13cc241
remove not needed variables
