+++ .github/workflows/ci.yaml
@@ -21,15 +21,6 @@ jobs:
         uses: actions/checkout@v3
         with:
           submodules: true
-      # For some reason doing this in pyproject.toml does not seem to work :(
-      - name: Set Windows compilation env vars
-        shell: bash
-        run: |
-          set -exo pipefail
-          echo "INCLUDE=$GITHUB_WORKSPACE\\src\\liblsl\\include" | tee -a $GITHUB_ENV
-          echo "LIB=$GITHUB_WORKSPACE\\src\\liblsl\\build\\Release" | tee -a $GITHUB_ENV
-          echo "ADD_PATH=$(cygpath -w $GITHUB_WORKSPACE)\\src\\liblsl\\build\\Release" | tee -a $GITHUB_ENV
-        if: runner.os == 'Windows'
       - uses: pypa/cibuildwheel@v2.21.3
       - uses: actions/upload-artifact@v4
         with:
