+++ .github/workflows/ci.yaml
@@ -80,9 +80,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        matrix:
-          mne-version: ["1.4.2", "1.5.0"]
-          python: ["3.10"]
+        mne-version: ["1.4.2", "1.5.0"]
+        python: ["3.10"]
     defaults:
       run:
         shell: bash
