### 2024-11-12 1e9f648
fix argument in editable install
