### 2024-11-12 5e200ea
update file inclusion/exclusion in sdist and bdist
