+++ MANIFEST.in
@@ -1,0 +1,26 @@
+# Include project files
+include LICENSE
+include pyproject.toml
+include README.md
+include setup.py
+
+# Include package source code
+graft src/mne_lsl
+graft src/liblsl
+
+# Prune package test files and liblsl configuration/doc files
+prune tests
+prune src/liblsl/docs
+prune src/liblsl/conda
+prune src/liblsl/examples
+prune src/liblsl/.github
+exclude src/liblsl/*.yml
+
+# Exclude compiled files and git files
+global-exclude *.py[cod] __pycache__ *.so *.dylib *.dll
+global-exclude .git*
+# Prune build artifacts if present
+prune build
+prune dist
+prune wheelhouse
+prune *.egg-info
