+++ pyproject.toml
@@ -194,14 +194,14 @@ ignore-decorators = ["mne_lsl.utils._docs.copy_doc", "qtpy.QtCore.Slot"]
 [tool.setuptools]
 include-package-data = true
 
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["mne_lsl*"]
-
 [tool.setuptools.package-data]
 'mne_lsl.datasets' = ['*registry.txt']
 "mne_lsl.lsl.lib" = ["*"]
 'mne_lsl.stream_viewer.control_gui.settings' = ['*.ini']
+
+[tool.setuptools.packages.find]
+include = ["mne_lsl*"]
+where = ["src"]
 
 [tool.tomlsort]
 all = true
