### 2024-11-12 afe5785
fix sdist
