+++ .github/workflows/ci.yaml
@@ -187,6 +187,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+        with:
+            submodules: true
       - uses: actions/setup-python@v5
         with:
           python-version: '3.13'
