### 2024-11-12 38f6a06
try again for editable install
