### 2024-11-12 cdd9324
fix editable test workflow
