### 2024-11-12 a5c6d1d
move doc to ci.yaml
