+++ .github/workflows/ci.yaml
@@ -184,9 +184,9 @@ jobs:
           token: ${{ secrets.CODECOV_TOKEN }}
 
   sdist:
-    timeout-minutes: 10
     name: create sdist
     runs-on: ubuntu-latest
+    timeout-minutes: 10
     steps:
       - uses: actions/checkout@v4
         with:
@@ -204,9 +204,9 @@ jobs:
 
   check:
     needs: [cibuildwheel, sdist]
-    timeout-minutes: 10
     name: run twine check
     runs-on: ubuntu-latest
+    timeout-minutes: 10
     steps:
       - uses: actions/checkout@v4
       - uses: actions/download-artifact@v4
@@ -222,17 +222,79 @@ jobs:
       - run: uv pip install twine --upgrade --quiet --system
       - run: twine check --strict dist/*
 
+  doc-build:
+    needs: cibuildwheel
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        shell: bash
+    steps:
+      - run: sudo apt update && sudo apt install -y binutils libpugixml-dev qtbase5-dev qt5-qmake optipng
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: '3.13'
+      - uses: astral-sh/setup-uv@v3
+      - uses: actions/download-artifact@v4
+        with:
+          pattern: cibw-wheels-ubuntu-latest-*
+          merge-multiple: true
+          path: dist
+      - run: ls -alt . && ls -alt dist/
+      - run: uv pip install mne-lsl[doc] --only-binary mne-lsl --find-links dist --pre --quiet --system
+      - run: mne-lsl sys-info --developer
+      - uses: ./.github/actions/get-testing-dataset
+        with:
+          sample: "true"
+          testing: "false"
+      - uses: ./.github/actions/retry-step
+        with:
+          command: "make -C doc html"
+          retry_error_codes: "2"
+          command_between_retries: "make -C doc clean"
+      - run: rm -R ./doc/_build/html/.doctrees
+      - uses: actions/upload-artifact@v4
+        with:
+          name: doc-dev
+          path: ./doc/_build/html
+
+  doc-deploy:
+    if: github.event_name == 'push'
+    needs: doc-build
+    name: deploy documentation
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+    permissions:
+      contents: write
+    defaults:
+      run:
+        shell: bash
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/download-artifact@v4
+        with:
+          name: doc-dev
+          path: ./doc-dev
+      - uses: JamesIves/github-pages-deploy-action@v4
+        with:
+          folder: ./doc-dev
+          target-folder: ./dev
+          git-config-name: 'github-actions[bot]'
+          git-config-email: 'github-actions[bot]@users.noreply.github.com'
+          single-commit: true
+          force: true
+
   publish:
-    needs: [check, test, test-mne-compat, test-pip-pre, test-editable]
+    if: github.event_name == 'release'
+    needs: [check, test, test-mne-compat, test-pip-pre, test-editable, doc-build]
     name: publish to PyPI
     runs-on: ubuntu-latest
+    timeout-minutes: 10
     permissions:
       id-token: write
     environment:
       name: pypi
       url: https://pypi.org/p/mne-lsl
-    timeout-minutes: 10
-    if: github.event_name == 'release'
     steps:
       - uses: actions/download-artifact@v4
         with:
