+++ .github/workflows/doc.yaml
@@ -1,80 +1,0 @@
-name: doc
-concurrency:
-  group: ${{ github.workflow }}-${{ github.event.number }}-${{ github.event.ref }}
-  cancel-in-progress: true
-on:  # yamllint disable-line rule:truthy
-  pull_request:
-  push:
-    branches: [main]
-  workflow_dispatch:
-
-jobs:
-  build:
-    timeout-minutes: 30
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        shell: bash
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-      - name: Setup Python 3.11
-        uses: actions/setup-python@v5
-        with:
-          python-version: 3.11
-      - name: Install system dependencies
-        uses: ./.github/actions/install-system-dependencies
-      - name: Install optipng (for sphinx-gallery)
-        run: sudo apt install optipng
-      - name: Install uv
-        uses: astral-sh/setup-uv@v3
-      - name: Install package
-        run: uv pip install --quiet --system .[doc]
-      - name: Display system information
-        run: mne-lsl sys-info --developer
-      - name: Get dataset
-        uses: ./.github/actions/get-testing-dataset
-        with:
-          sample: "true"
-          testing: "false"
-      - name: Build doc
-        uses: ./.github/actions/retry-step
-        with:
-          command: "make -C doc html"
-          retry_error_codes: "2"
-          command_between_retries: "make -C doc clean"
-      - name: Prune sphinx environment
-        run: rm -R ./doc/_build/html/.doctrees
-      - name: Upload documentation
-        uses: actions/upload-artifact@v4
-        with:
-          name: doc-dev
-          path: ./doc/_build/html
-
-  deploy:
-    if: github.event_name == 'push'
-    needs: build
-    timeout-minutes: 10
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-    defaults:
-      run:
-        shell: bash
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-      - name: Download documentation
-        uses: actions/download-artifact@v4
-        with:
-          name: doc-dev
-          path: ./doc-dev
-      - name: Deploy
-        uses: JamesIves/github-pages-deploy-action@v4
-        with:
-          folder: ./doc-dev
-          target-folder: ./dev
-          git-config-name: 'github-actions[bot]'
-          git-config-email: 'github-actions[bot]@users.noreply.github.com'
-          single-commit: true
-          force: true
