+++ .github/workflows/ci.yaml
@@ -190,7 +190,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
         with:
-            submodules: true
+          submodules: true
       - uses: actions/setup-python@v5
         with:
           python-version: '3.13'
