+++ .github/workflows/stubs.yaml
@@ -17,22 +17,16 @@ jobs:
         shell: bash
     steps:
       - run: sudo apt update && sudo apt install -y binutils libpugixml-dev
-      - name: Checkout repository
-        uses: actions/checkout@v4
+      - uses: actions/checkout@v4
         with:
           submodules: true
-      - name: Setup Python 3.13
-        uses: actions/setup-python@v5
+      - uses: actions/setup-python@v5
         with:
           python-version: 3.13
-      - name: Install uv
-        uses: astral-sh/setup-uv@v3
-      - name: Install package
-        run: uv pip install --quiet --system -e .[stubs]
-      - name: Display system information
-        run: mne-lsl sys-info --developer
-      - name: Generate stub files
-        run: python tools/stubgen.py
+      - uses: astral-sh/setup-uv@v3
+      - run: uv pip install -e .[stubs] --quiet --system
+      - run: mne-lsl sys-info --developer
+      - run: python tools/stubgen.py
       - name: Push stub files
         if: github.event_name == 'schedule' || github.event_name == 'workflow_dispatch'
         env:
