### 2024-11-12 fa01265
use checkout@v4 to fix deprecation
