+++ .github/workflows/ci.yaml
@@ -20,6 +20,23 @@ jobs:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-13, macos-latest]
         arch: [native]
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          submodules: true
+      - uses: pypa/cibuildwheel@v2.21.3
+      - uses: actions/upload-artifact@v4
+        with:
+          name: cibw-wheels-${{ matrix.os }}-${{ matrix.arch }}-${{ strategy.job-index }}
+          path: ./wheelhouse/*.whl
+
+  cibuildwheel_emulated:
+    name: build wheels on ${{ matrix.os }} ${{ matrix.arch }}
+    runs-on: ${{ matrix.os }}
+    timeout-minutes: 20
+    strategy:
+      fail-fast: false
+      matrix:
         include:
           - os: ubuntu-latest
             arch: aarch64
@@ -27,17 +44,22 @@ jobs:
       - uses: actions/checkout@v4
         with:
           submodules: true
+      - name: Triage aarch64 support
+        run: bash ./tools/triage_build.sh "${{ matrix.arch }}" "${{ github.event.pull_request.head.sha || github.sha }}"
+        id: triage
       - uses: docker/setup-qemu-action@v3
         with:
-          platforms: aarch64
-        if: runner.os == 'Linux' && matrix.arch == 'aarch64'
+          platforms: ${{ matrix.arch }}
+        if: steps.triage.outputs.skip != '1'
       - uses: pypa/cibuildwheel@v2.21.3
         env:
           CIBW_ARCHS: ${{ matrix.arch }}
+        if: steps.triage.outputs.skip != '1'
       - uses: actions/upload-artifact@v4
         with:
-          name: cibw-wheels-${{ matrix.os }}-${{ strategy.job-index }}
+          name: cibw-wheels-${{ matrix.os }}-${{ matrix.arch }}-${{ strategy.job-index }}
           path: ./wheelhouse/*.whl
+        if: steps.triage.outputs.skip != '1'
 
   test:
     needs: cibuildwheel
@@ -212,7 +234,7 @@ jobs:
           path: ./dist/*.tar.gz
 
   check:
-    needs: [cibuildwheel, sdist]
+    needs: [cibuildwheel, cibuildwheel_emulated, sdist]
     name: run twine check
     runs-on: ubuntu-latest
     timeout-minutes: 10
