+++ .github/workflows/ci.yaml
@@ -30,7 +30,22 @@ jobs:
           name: cibw-wheels-${{ matrix.os }}-${{ matrix.arch }}-${{ strategy.job-index }}
           path: ./wheelhouse/*.whl
 
+  get_commit_message:
+    runs-on: ubuntu-latest
+    outputs:
+      commit_message: ${{ steps.get_message.outputs.commit_message }}
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          # ensure the correct commit is checked out for PRs or pushes
+          ref: ${{ github.event.pull_request.head.sha || github.sha }}
+          fetch-depth: 1
+      - run: echo "commit_message=$(git log -1 --pretty=%B)" >> $GITHUB_OUTPUT
+        id: get_message
+
   cibuildwheel_emulated:
+    needs: get_commit_message
+    if: ${{ github.event_name != 'pull_request' || contains(needs.get_commit_message.outputs.commit_message, '[aarch64]') }}
     name: build wheels on ${{ matrix.os }} ${{ matrix.arch }}
     runs-on: ${{ matrix.os }}
     timeout-minutes: 20
@@ -43,24 +58,17 @@ jobs:
     steps:
       - uses: actions/checkout@v4
         with:
-          fetch-depth: 0  # get the non-merge commit for PRs
           submodules: true
-      - name: Triage aarch64 support
-        run: bash ./tools/triage_build.sh "${{ matrix.arch }}" "${{ github.event.pull_request.head.sha || github.sha }}"
-        id: triage
       - uses: docker/setup-qemu-action@v3
         with:
           platforms: ${{ matrix.arch }}
-        if: steps.triage.outputs.skip != '1'
       - uses: pypa/cibuildwheel@v2.21.3
         env:
           CIBW_ARCHS: ${{ matrix.arch }}
-        if: steps.triage.outputs.skip != '1'
       - uses: actions/upload-artifact@v4
         with:
           name: cibw-wheels-${{ matrix.os }}-${{ matrix.arch }}-${{ strategy.job-index }}
           path: ./wheelhouse/*.whl
-        if: steps.triage.outputs.skip != '1'
 
   test:
     needs: cibuildwheel
@@ -235,6 +243,8 @@ jobs:
           path: ./dist/*.tar.gz
 
   check:
+    # run even if 'cibuildwheel_emulated' is skipped
+    if: ${{ always() }}
     needs: [cibuildwheel, cibuildwheel_emulated, sdist]
     name: run twine check
     runs-on: ubuntu-latest
