### 2024-11-13 599f134
remove unnecessary step [aarch64]
