### 2024-11-13 4f7032a
add environment variable to skip build process
