### 2024-11-13 5da48ce
improve handling of environment variables
