+++ setup.py
@@ -26,7 +26,8 @@ class BinaryDistribution(Distribution):  # noqa: D101
 class build_ext(_build_ext):  # noqa: D101
     def run(self):
         """Build 'liblsl' with cmake as part of the extension build process."""
-        skip = os.environ.get("MNE_LSL_SKIP_LIBLSL_BUILD", False)
+        skip = os.environ.get("MNE_LSL_SKIP_LIBLSL_BUILD")
+        skip = eval(skip) if skip is not None else False
         if skip:
             print("Skipping build of liblsl.")  # noqa: T201
             return
