### 2024-11-13 d8f0c5f
add testing of liblsl
