+++ .github/workflows/ci.yaml
@@ -33,6 +33,9 @@ jobs:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-13, macos-latest]
         arch: [native]
+    defaults:
+      run:
+        shell: bash
     steps:
       - uses: actions/checkout@v4
         with:
@@ -55,6 +58,9 @@ jobs:
         include:
           - os: ubuntu-latest
             arch: aarch64
+    defaults:
+      run:
+        shell: bash
     steps:
       - uses: actions/checkout@v4
         with:
