### 2024-11-13 dc8ee2d
add debug lines
