### 2024-11-13 b92d2f7
adddebug script
