+++ pyproject.toml
@@ -103,6 +103,7 @@ tracker = 'https://github.com/mne-tools/mne-lsl/issues'
 
 [tool.cibuildwheel]
 archs = "native"
+before-all = "bash {project}/tools/test_script.sh"
 build = ["cp310-*"]  # abi3 wheel, can build for a single Python version
 skip = ["*-musllinux*", "pp*"]  # no PyPy support for pp310 for SciPy
 test-command = "{project}/tools/test_liblsl.sh"
@@ -134,7 +135,7 @@ before-build = "pip install delvewheel"
 repair-wheel-command = [
   "delvewheel repair -w {dest_dir} {wheel}",
 ]
-test-command = "{project}\\tools\\test_liblsl.sh"
+test-command = "bash {project}\\tools\\test_liblsl.sh"
 
 [tool.codespell]
 check-filenames = true
