+++ tools/test_script.sh
@@ -1,1 +1,0 @@
-echo "Test"
