+++ .github/workflows/ci.yaml
@@ -41,6 +41,7 @@ jobs:
       - uses: actions/checkout@v4
         with:
           submodules: true
+      - uses: astral-sh/setup-uv@v3
       - uses: pypa/cibuildwheel@v2.21.3
       - uses: actions/upload-artifact@v4
         with:
@@ -69,6 +70,7 @@ jobs:
       - uses: docker/setup-qemu-action@v3
         with:
           platforms: ${{ matrix.arch }}
+      - uses: astral-sh/setup-uv@v3
       - uses: pypa/cibuildwheel@v2.21.3
         env:
           CIBW_ARCHS: ${{ matrix.arch }}
