### 2024-11-13 0905d3c
add artifact upload and disable liblsl for aarch64 emulation
