### 2024-11-13 ebb682d
set env variable based on commit message
