+++ .github/workflows/ci.yaml
@@ -26,6 +26,7 @@ jobs:
         id: get_message
 
   cibuildwheel:
+    needs: get_commit_message
     name: build wheels on ${{ matrix.os }} ${{ matrix.arch }}
     runs-on: ${{ matrix.os }}
     timeout-minutes: 10
@@ -43,6 +44,8 @@ jobs:
           submodules: true
       - uses: astral-sh/setup-uv@v3
       - uses: pypa/cibuildwheel@v2.21.3
+        env:
+          MNE_LSL_LIBLSL_BUILD_UNITTESTS: ${{ contains(needs.get_commit_message.outputs.commit_message, '[test-liblsl]') && '1' || '0' }}
       - uses: actions/upload-artifact@v4
         with:
           name: cibw-wheels-${{ matrix.os }}-${{ matrix.arch }}-${{ strategy.job-index }}
