### 2024-11-14 60ca6f8
increase tolerance and test commit message tag [liblsl]
