### 2024-11-14 582b81a
ENH: Build and bundle liblsl in wheels (#351)
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@fcbg.ch>
