+++ tests/conftest.py
@@ -65,6 +65,8 @@ def pytest_configure(config: pytest.Config) -> None:
     ignore:Python 3.14 will, by default.*:DeprecationWarning
     # Matplotlib deprecation issued in VSCode test debugger
     ignore:.*interactive_bk.*:matplotlib._api.deprecation.MatplotlibDeprecationWarning
+    # tkinter
+    ignore:Exception ignored in.*__del__.*:pytest.PytestUnraisableExceptionWarning
     """
     for warning_line in warning_lines.split("\n"):
         warning_line = warning_line.strip()
