+++ mne_lsl/lsl/load_liblsl.py
@@ -1,647 +1,0 @@
-from __future__ import annotations
-
-import os
-import platform
-import subprocess
-import tarfile
-import zipfile
-from ctypes import CDLL, c_char_p, c_double, c_long, c_void_p, sizeof
-from ctypes.util import find_library
-from pathlib import Path
-from shutil import move, rmtree
-from typing import TYPE_CHECKING
-
-import pooch
-import requests
-from mne import get_config, set_config
-
-from .._version import __version__
-from ..utils._checks import ensure_path
-from ..utils._path import walk
-from ..utils.logs import logger, warn
-
-if TYPE_CHECKING:
-    from pooch import Pooch
-
-
-# minimum liblsl version. The major version is given by version // 100
-# and the minor version is given by version % 100.
-_VERSION_MIN: int = 115
-# liblsl objects created with the same protocol version are inter-compatible.
-_VERSION_PROTOCOL: int = 110  # noqa: W0612
-_PLATFORM: str = platform.system().lower().strip()
-_PLATFORM_SUFFIXES: dict[str, str] = {
-    "windows": ".dll",
-    "darwin": ".dylib",
-    "linux": ".so",
-}
-# generic error message
-_ERROR_MSG: str = (
-    "Please visit LIBLSL library github page (https://github.com/sccn/liblsl) and "
-    "install a release in the system directories or provide its path in the "
-    "environment variable MNE_LSL_LIB or PYLSL_LIB."
-)
-
-
-def load_liblsl() -> CDLL:
-    """Load the binary LSL library on the system.
-
-    The library is loaded in the following order:
-
-    1. Search in the environment variables.
-    2. Search in the system folder.
-    3. Search in the defined library folder.
-    4. Fetch on GitHub.
-    """
-    if _PLATFORM not in _PLATFORM_SUFFIXES:  # pragma: no cover
-        raise RuntimeError(
-            "The OS could not be determined. Please open an issue on GitHub and "
-            "provide the error traceback to the developers."
-        )
-    libpath = _load_liblsl_environment_variables()
-    libpath = _load_liblsl_system() if libpath is None else libpath
-    libpath = _load_liblsl_mne_lsl() if libpath is None else libpath
-    libpath = _fetch_liblsl() if libpath is None else libpath
-    assert isinstance(libpath, str)  # sanity-check
-    lib = CDLL(libpath)
-    return _set_types(lib)
-
-
-def _load_liblsl_environment_variables() -> str | None:
-    """Load the binary LSL library from the environment variables.
-
-    Returns
-    -------
-    libpath : str | None
-        Path to the binary LSL library or None if it could not be found.
-    """
-    for variable in ("MNE_LSL_LIB", "PYLSL_LIB"):
-        libpath = os.environ.get(variable, None)
-        if libpath is None:
-            logger.debug("The environment variable '%s' is not set.", variable)
-            continue
-        logger.debug(
-            "Attempting to load libpath '%s' stored in the environment variable '%s'.",
-            libpath,
-            variable,
-        )
-        # even if the path is not valid, we still try to load it and issue a second
-        # generic warning 'can not be loaded' if it fails.
-        _is_valid_libpath(libpath)
-        libpath, version = _attempt_load_liblsl(libpath)
-        if version is None:
-            continue
-        # we do not accept outdated versions from the environment variables.
-        if _is_valid_version(libpath, version):
-            return libpath
-    return None
-
-
-def _load_liblsl_system() -> str | None:
-    """Load the binary LSL library from the system path/folders.
-
-    Returns
-    -------
-    libpath : str | None
-        Path to the binary LSL library or None if it could not be found.
-    """
-    libpath = find_library("lsl")
-    if libpath is None:
[…]
