+++ src/mne_lsl/lsl/load_liblsl.py
@@ -1,0 +1,232 @@
+from __future__ import annotations
+
+import os
+import platform
+from ctypes import CDLL, c_char_p, c_double, c_long, c_void_p
+from importlib.resources import files
+
+from ..utils.logs import logger, warn
+
+# minimum recommended liblsl version. The major version is given by version // 100
+# and the minor version is given by version % 100.
+_VERSION_MIN: int = 115
+# liblsl objects created with the same protocol version are inter-compatible.
+_VERSION_PROTOCOL: int = 110  # noqa: W0612
+
+
+def load_liblsl() -> CDLL:
+    """Load the binary LSL library on the system.
+
+    The library is loaded in the following order:
+
+    1. Search in the environment variables.
+    2. Search in the system folder.
+    3. Search in the defined library folder.
+    4. Fetch on GitHub.
+    """
+    libpath = _load_liblsl_environment_variables()
+    libpath = _load_liblsl_wheel_path()
+    assert isinstance(libpath, str)  # sanity-check
+    lib = CDLL(libpath)
+    _set_types(lib)
+    return lib
+
+
+def _load_liblsl_environment_variables(
+    *, version_min: int = _VERSION_MIN
+) -> str | None:
+    """Load the binary LSL library from the environment variables.
+
+    Parameters
+    ----------
+    version_min : int
+        Minimum version of the LSL library.
+
+    Returns
+    -------
+    libpath : str | None
+        Path to the binary LSL library or None if it could not be found.
+    """
+    for variable in ("MNE_LSL_LIB", "PYLSL_LIB"):
+        libpath = os.environ.get(variable, None)
+        if libpath is None:
+            logger.debug("The environment variable '%s' is not set.", variable)
+            continue
+        logger.debug(
+            "Attempting to load libpath '%s' stored in the environment variable '%s'.",
+            libpath,
+            variable,
+        )
+        try:
+            lib = CDLL(libpath)
+            version = lib.lsl_library_version()
+        except OSError:
+            continue
+        # we warn for outdated versions
+        if version < version_min:
+            warn(
+                f"The LIBLSL '{libpath}' (version {version // 100}.{version % 100}) is "
+                "outdated, use at your own discretion. MNE-LSL recommends to use "
+                f"version {version_min // 100}.{version_min % 100} and above."
+            )
+        return libpath
+    return None
+
+
+def _load_liblsl_wheel_path() -> str:
+    """Load the binary LSL library from the wheel path.
+
+    Returns
+    -------
+    libpath : str
+        Path to the binary LSL library bundled with mne-lsl.
+    """
+    patterns: dict[str, str] = {
+        "Linux": "liblsl*.so*",
+        "Darwin": "liblsl*.dylib",
+        "Windows": "lsl*.dll",
+    }
+    pattern = patterns.get(platform.system(), None)
+    if pattern is None:  # pragma: no cover
+        raise RuntimeError(
+            f"Unsupported platform {platform.system()}. Please use the environment "
+            "variable MNE_LSL_LIB or PYLSL_LIB to provide the path to LIBLSL."
+        )
+    folder = files("mne_lsl.lsl") / "lib"
[…]
