+++ .github/workflows/publish.yaml
@@ -1,35 +1,0 @@
-name: publish
-
-on:  # yamllint disable-line rule:truthy
-  release:
-    types: [published]
-
-jobs:
-  pypi:
-    timeout-minutes: 10
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-      - name: Setup Python 3.11
-        uses: actions/setup-python@v5
-        with:
-          python-version: 3.11
-      - name: Install system dependencies
-        uses: ./.github/actions/install-system-dependencies
-      - name: Install uv
-        uses: astral-sh/setup-uv@v3
-      - name: Install package
-        run: uv pip install --quiet --system -e .[build,stubs]
-      - name: Display system information
-        run: mne-lsl sys-info --developer
-      - name: Generate stub files
-        run: python tools/stubgen.py
-      - name: Build and publish
-        env:
-          TWINE_USERNAME: __token__
-          TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN  }}
-        run: |
-          python -m build
-          twine check dist/*
-          twine upload dist/*

[…]
