### 2024-11-14 aa91e4b
Fix parsing of commit message: select only the commit subject (#355)
