+++ doc/development/changes/latest.rst
@@ -19,5 +19,4 @@
 Version 1.7
 ===========
 
-- Add the MNE configuration variable ``MNE_LSL_LIB_FOLDER`` to set the liblsl fetcher download directory (:pr:`350` by `Mathieu Scheltienne`_)
-- Change default liblsl fetcher download directory to ``MNE_DATA / MNE-LSL-data / liblsl`` (:pr:`350` by `Mathieu Scheltienne`_)
+- Package ``liblsl`` with ``mne-lsl`` in wheels (:pr:`351` by `Eric Larson`_ and `Mathieu Scheltienne`_)
