### 2024-11-15 a53064b
add workflow to simplify release process
