### 2024-11-15 5426e94
Add workflow to simplify release process (#357)
