### 2024-11-15 f29f4f9
Remove outdated and unused dependencies (#358)
