### 2024-11-17 a998442
fix numerical approximation error in example
