### 2024-11-18 ee64445
Bump codecov/codecov-action from 4 to 5 (#360)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
