### 2024-11-18 e7cbf67
set error handling in bash script to test liblsl
