### 2024-11-21 7a24c6b
fix more type-hints
