### 2024-11-21 3fab197
fix workflow tag parsing
