### 2024-11-21 14ae224
fix workflow by running pypi publication after updating release assets
