+++ .github/workflows/ci.yaml
@@ -402,8 +402,8 @@ jobs:
           pattern: cibw-wheels-*
           merge-multiple: true
           path: dist
-      - uses: pypa/gh-action-pypi-publish@release/v1
       - uses: softprops/action-gh-release@v2
         with:
           files: dist/*
           tag_name: ${{ github.event.release.tag_name }}
+      - uses: pypa/gh-action-pypi-publish@release/v1
