+++ doc/development/changes/1.2.rst
@@ -14,8 +14,8 @@
 
 .. include:: ./authors.inc
 
-Version 1.2
-===========
+Version 1.2.0
+=============
 
 - Apply MNE design philosophy by returning ``self`` in the methods modifying a :class:`~mne_lsl.stream.StreamLSL` and :class:`~mne_lsl.player.PlayerLSL` in-place (:pr:`200` by `Mathieu Scheltienne`_)
 - Add argument ``annotations`` to :class:`~mne_lsl.player.PlayerLSL` to stream annotations on a separate :class:`~mne_lsl.lsl.StreamOutlet` (:pr:`202` by `Mathieu Scheltienne`_)
