### 2024-11-22 c934365
Fix loading of liblsl from environment variable (#363)
