+++ doc/development/changes/1.3.rst
@@ -14,8 +14,8 @@
 
 .. include:: ./authors.inc
 
-Version 1.3
-===========
+Version 1.3.0
+=============
 
 - Improve logging to set MNE's logger level to currently set MNE-LSL level in MNE-LSL functions using MNE's function under the hood (:pr:`219` by `Mathieu Scheltienne`_)
 - Use the currently set logger level as replacement for ``verbose=None`` in MNE-LSL functions (:pr:`219` by `Mathieu Scheltienne`_)
@@ -26,3 +26,8 @@ API changes (backward incompatible)
 -----------------------------------
 
 - Add ``n_repeat`` argument to :class:`~mne_lsl.player.PlayerLSL` to support looping a file a fix number of times (:pr:`223` by `Mathieu Scheltienne`_)
+
+Version 1.3.1
+=============
+
+- Improve handling of warnings through the codebase (:pr:`243` by `Mathieu Scheltienne`_)
