### 2024-11-22 71a4ee6
Fix loading of liblsl from environment variable (#363)
