+++ doc/development/changes/1.5.rst
@@ -14,8 +14,8 @@
 
 .. include:: ./authors.inc
 
-Version 1.5
-===========
+Version 1.5.0
+=============
 
 - Add ``exclude='bads'`` argument to :meth:`~mne_lsl.stream.StreamLSL.get_data` to control exclusion from ``picks`` (:pr:`258` by `Mathieu Scheltienne`_)
 - Add :class:`~mne_lsl.stream.EpochsStream` to epoch a Stream object object on the fly (:pr:`258` by `Mathieu Scheltienne`_)
