+++ doc/development/changes/1.6.rst
@@ -14,12 +14,16 @@
 
 .. include:: ./authors.inc
 
-Version 1.6
-===========
+Version 1.6.0
+=============
 
 - Add JOSS draft and workflow to generate the PDF (:pr:`316` by `Mathieu Scheltienne`_)
 - Improve disconnection of :class:`~mne_lsl.stream.EpochsStream` especially when an exception is raised during initialization (:pr:`317` by `Mathieu Scheltienne`_)
 - Improve total code coverage (:pr:`317` by `Mathieu Scheltienne`_)
 - Add mailmap to the repository (:pr:`327` by `Daniel McCloy`_)
 - Fix indexing when replaying datasets from non-preloaded :class:`~mne.io.Raw` (:pr:`330` by `Toni M. Brotons`_)
+
+Version 1.6.1
+=============
+
 - Safer parse of channel names in the legacy stream viewer (by `Mathieu Scheltienne`_)
