### 2024-11-22 7cedb4a
fix workflow for release on maint branch
