+++ .github/workflows/ci.yaml
@@ -328,8 +328,6 @@ jobs:
     runs-on: ubuntu-latest
     permissions:
       contents: write
-    env:
-      RELEASE_TAG: ${{ github.event.release.tag_name }}
     steps:
       - uses: actions/checkout@v4
         with:
@@ -339,16 +337,7 @@ jobs:
         with:
           name: doc-dev
           path: ./doc-dev
-      - name: Move doc to release folder
-        if: github.event.release.target_commitish != 'main'
-        run: |
-          if [ -d "$RELEASE_TAG" ]; then
-            echo "Folder $RELEASE_TAG already exists. Exiting."
-            exit 1
-          fi
-          mv doc-dev "$RELEASE_TAG"
       - name: Get previous release tag
-        if: github.event.release.target_commitish == 'main'
         run: |
           PREVIOUS_RELEASE_TAG=$(git tag --sort=-creatordate | sed "/^$RELEASE_TAG$/d" | sed -n 1p)
           if [ -z "$PREVIOUS_RELEASE_TAG" ]; then
@@ -357,8 +346,9 @@ jobs:
           fi
           echo "Previous release tag: $PREVIOUS_RELEASE_TAG"
           echo "PREVIOUS_RELEASE_TAG=$PREVIOUS_RELEASE_TAG" >> $GITHUB_ENV
+        env:
+          RELEASE_TAG: ${{ github.event.release.tag_name }}
       - name: Move stable to previous release tag
-        if: github.event.release.target_commitish == 'main'
         run: |
           if [ -d "stable" ]; then
             if [ -d "$PREVIOUS_RELEASE_TAG" ]; then
@@ -370,12 +360,9 @@ jobs:
             echo "No stable folder found."
             exit 1
           fi
-      - name: Copy built docs to stable and dev
+      - run: mv doc-dev stable
+      - run: rm -rf dev && cp -r stable dev
         if: github.event.release.target_commitish == 'main'
-        run: |
-          rm -rf dev
-          mv doc-dev stable
-          cp -r stable dev
       - name: Commit and push changes
         run: |
           git config user.name 'github-actions[bot]'
