### 2024-11-23 2cb9866
fix stub gen [skip ci]
