+++ .github/workflows/stubs.yaml
@@ -37,7 +37,7 @@ jobs:
           if [ -n "$(git status --porcelain)" ]; then
             branch_name="update-stub-files-$(date +'%Y%m%d-%H%M%S')"
             git checkout -b "$branch_name"
-            git add mne_lsl/\*.pyi
+            git add src/mne_lsl/\*.pyi
             git commit -m "deploy stub files [ci skip]"
             git push --set-upstream origin "$branch_name"
             gh pr create --title "Update stub files" --body "Auto-generated PR for updating stub files." --base main
