+++ doc/development/changes/1.6.rst
@@ -14,6 +14,11 @@
 
 .. include:: ./authors.inc
 
+Version 1.6.1
+=============
+
+- Safer parse of channel names in the legacy stream viewer (by `Mathieu Scheltienne`_)
+
 Version 1.6.0
 =============
 
@@ -22,8 +27,3 @@ Version 1.6.0
 - Improve total code coverage (:pr:`317` by `Mathieu Scheltienne`_)
 - Add mailmap to the repository (:pr:`327` by `Daniel McCloy`_)
 - Fix indexing when replaying datasets from non-preloaded :class:`~mne.io.Raw` (:pr:`330` by `Toni M. Brotons`_)
-
-Version 1.6.1
-=============
-
-- Safer parse of channel names in the legacy stream viewer (by `Mathieu Scheltienne`_)
