+++ doc/development/changes/1.3.rst
@@ -14,6 +14,11 @@
 
 .. include:: ./authors.inc
 
+Version 1.3.1
+=============
+
+- Improve handling of warnings through the codebase (:pr:`243` by `Mathieu Scheltienne`_)
+
 Version 1.3.0
 =============
 
@@ -26,8 +31,3 @@ API changes (backward incompatible)
 -----------------------------------
 
 - Add ``n_repeat`` argument to :class:`~mne_lsl.player.PlayerLSL` to support looping a file a fix number of times (:pr:`223` by `Mathieu Scheltienne`_)
-
-Version 1.3.1
-=============
-
-- Improve handling of warnings through the codebase (:pr:`243` by `Mathieu Scheltienne`_)
