### 2024-11-25 c8c1ee2
Bump astral-sh/setup-uv from 3 to 4 (#367)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
