### 2024-11-29 463951b
remove unnecessary copy
