### 2024-11-29 bb7af33
better docstrings
