### 2024-11-29 a59c84e
Don't use public get_data in epoch acquisition to avoid resetting the number of new samples (#371)
