### 2024-11-29 39d1b51
Deprecate acquisition_delay=0 in favor of None (#372)
