### 2024-11-29 c60a156
Prepare for 1.8 release (#373)
