+++ doc/development/changes/index.rst
@@ -4,6 +4,7 @@ Changelog
 .. toctree::
     :titlesonly:
 
+    latest.rst
     1.8.rst
     1.7.rst
     1.6.rst
