### 2024-11-29 2550a40
remove deprecation
