+++ src/mne_lsl/stream/base.py
@@ -291,20 +291,14 @@ class BaseStream(ABC, ContainsMixin, SetChannelsMixin):
             return self
         if acquisition_delay is not None:
             check_type(acquisition_delay, ("numeric",), "acquisition_delay")
-            if acquisition_delay < 0:
+            if acquisition_delay <= 0:
                 raise ValueError(
-                    "The acquisition delay must be a positive number defining the "
-                    "delay at which new samples are acquired in seconds. For instance, "
-                    "0.2 corresponds to a pull every 200 ms. None corresponds to "
-                    f"manual acquisition. The provided {acquisition_delay} is invalid."
+                    "The acquisition delay must be a strictly positive number defining "
+                    "the delay at which new samples are acquired in seconds. For "
+                    "instance, 0.2 corresponds to a pull every 200 ms. None "
+                    "corresponds to manual acquisition. The provided "
+                    f"{acquisition_delay} is invalid."
                 )
-            if acquisition_delay == 0:
-                warn(
-                    "Argument acquisition_delay=0 is deprecated in favor of "
-                    "acquisition_delay=None.",
-                    DeprecationWarning,
-                )
-                acquisition_delay = None
         self._acquisition_delay = acquisition_delay
         self._n_new_samples = 0
         self._executor = (
