### 2024-11-29 f4f0f73
Post release 1.8 and handle deprecation (#374)
