### 2024-11-30 14addef
add stub generation via autofix.ci
