+++ .github/workflows/ci.yaml
@@ -30,14 +30,14 @@ jobs:
     runs-on: ubuntu-latest
     timeout-minutes: 10
     steps:
-    - uses: actions/checkout@v4
-    - uses: actions/setup-python@v5
-      with:
-        python-version: 3.13
-    - uses: astral-sh/setup-uv@v4
-    - run: uv pip install -e .[stubs] --quiet --system
-    - run: python tools/stubgen.py
-    - uses: autofix-ci/action@v1.3.1
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: 3.13
+      - uses: astral-sh/setup-uv@v4
+      - run: uv pip install -e .[stubs] --quiet --system
+      - run: python tools/stubgen.py
+      - uses: autofix-ci/action@v1.3.1
 
   cibuildwheel:
     needs: get_commit_message
