+++ .github/workflows/stubs.yaml
@@ -12,11 +12,7 @@ jobs:
   generate:
     timeout-minutes: 10
     runs-on: ubuntu-latest
-    defaults:
-      run:
-        shell: bash
     steps:
-      - run: sudo apt update && sudo apt install -y binutils libpugixml-dev
       - uses: actions/checkout@v4
         with:
           submodules: true
@@ -27,6 +23,8 @@ jobs:
       - run: uv pip install -e .[stubs] --quiet --system
       - run: mne-lsl sys-info --developer
       - run: python tools/stubgen.py
+      - uses: autofix-ci/action@v1.3.1
+        if: ${{ github.event_name == 'pull_request' }}
       - name: Push stub files
         if: github.event_name == 'schedule' || github.event_name == 'workflow_dispatch'
         env:
