+++ .github/workflows/stubs.yaml
@@ -26,7 +26,7 @@ jobs:
       - uses: autofix-ci/action@v1.3.1
         if: ${{ github.event_name == 'pull_request' }}
       - name: Push stub files
-        if: github.event_name == 'schedule' || github.event_name == 'workflow_dispatch'
+        if: ${{ github.event_name == 'schedule' || github.event_name == 'workflow_dispatch' }}
         env:
           GH_TOKEN: ${{ github.token }}
         run: |
