+++ .github/workflows/draft-pdf.yml
@@ -13,16 +13,13 @@ jobs:
     runs-on: ubuntu-latest
     name: draft PDF
     steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@master
+      - uses: actions/checkout@v4
+      - uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
           # This should be the path to the paper within your repo.
           paper-path: paper/paper.md
-      - name: Upload
-        uses: actions/upload-artifact@v4
+      - uses: actions/upload-artifact@v4
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled PDF. Note, this
