### 2024-12-02 1a4c1a5
Add autofix.ci to run stub file generation on pull request (#376)
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
