### 2024-12-10 a85e83c
[pre-commit.ci] pre-commit autoupdate (#379)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
