### 2024-12-16 dffcbba
Remove pylsl comparison workflow (#381)
