+++ .github/workflows/pylsl.yaml
@@ -1,32 +1,0 @@
-name: pylsl
-concurrency:
-  group: ${{ github.workflow }}-${{ github.event.number }}-${{ github.event.ref }}
-  cancel-in-progress: true
-on:  # yamllint disable-line rule:truthy
-  workflow_dispatch:
-  schedule:
-    - cron: '0 8 * * 1'
-
-jobs:
-  sha256:
-    timeout-minutes: 2
-    runs-on: ubuntu-latest
-    env:
-      SHA256_PYLSL: 0a384000acdd5f831376d142b4aea373693f1e356ab1cd3dbdef10896acf74e3
-    steps:
-      - name: Checkout pylsl repository
-        uses: actions/checkout@v4
-        with:
-          repository: labstreaminglayer/pylsl
-      - name: Compare checksum
-        run: echo "$SHA256_PYLSL pylsl/pylsl.py" | sha256sum -c
-      - name: Create GitHub issue
-        if: ${{ failure() }}
-        uses: dacbd/create-issue-action@main
-        with:
-          token: ${{ github.token }}
-          title: pyLSL is out of sync with MNE-LSL
-          body: |
-            Please visit https://github.com/labstreaminglayer/pylsl and review the changes.
-            [Failed Run](https://github.com/${{ github.repository }}/actions/runs/${{ github.run_id }})
-          assignees: mscheltienne
