+++ 1.6.0/_sources/sg_execution_times.rst.txt
@@ -1,0 +1,64 @@
+
+:orphan:
+
+.. _sphx_glr_sg_execution_times:
+
+
+Computation times
+=================
+**05:20.988** total execution time for 10 files **from all galleries**:
+
+.. container::
+
+  .. raw:: html
+
+    <style scoped>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
+    </style>
+    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
+    <script type="text/javascript" class="init">
+    $(document).ready( function () {
+        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
+    } );
+    </script>
+
+  .. list-table::
+   :header-rows: 1
+   :class: table table-striped sg-datatable
+
+   * - Example
+     - Time
+     - Mem (MB)
+   * - :ref:`sphx_glr_generated_examples_40_decode.py` (``../examples/40_decode.py``)
+     - 01:19.441
+     - 632.9
+   * - :ref:`sphx_glr_generated_examples_30_real_time_evoked_responses.py` (``../examples/30_real_time_evoked_responses.py``)
+     - 00:59.993
+     - 606.4
+   * - :ref:`sphx_glr_generated_tutorials_40_epochs.py` (``../tutorials/40_epochs.py``)
+     - 00:52.696
+     - 721.6
+   * - :ref:`sphx_glr_generated_examples_20_bandpower.py` (``../examples/20_bandpower.py``)
+     - 00:49.017
+     - 203.9
+   * - :ref:`sphx_glr_generated_examples_10_peak_detection.py` (``../examples/10_peak_detection.py``)
+     - 00:38.301
+     - 191.6
+   * - :ref:`sphx_glr_generated_tutorials_30_stream_manual.py` (``../tutorials/30_stream_manual.py``)
+     - 00:12.609
+     - 277.0
+   * - :ref:`sphx_glr_generated_tutorials_20_player_annotations.py` (``../tutorials/20_player_annotations.py``)
+     - 00:10.281
+     - 203.0
+   * - :ref:`sphx_glr_generated_examples_00_player_separate_process.py` (``../examples/00_player_separate_process.py``)
+     - 00:09.752
+     - 187.9
+   * - :ref:`sphx_glr_generated_tutorials_00_introduction.py` (``../tutorials/00_introduction.py``)
+     - 00:05.536
+     - 297.4
+   * - :ref:`sphx_glr_generated_tutorials_10_low_level_API.py` (``../tutorials/10_low_level_API.py``)
+     - 00:03.361
+     - 201.2
