+++ 1.3.1/_sources/changes/1.3.rst.txt
@@ -1,0 +1,35 @@
+.. NOTE: we use cross-references to highlight new functions and classes.
+   Please follow the examples below, so the changelog page will have a link to
+   the function/class documentation.
+
+.. NOTE: there are 3 separate sections for changes, based on type:
+   - "Enhancements" for new features
+   - "Bugs" for bug fixes
+   - "API changes" for backward-incompatible changes
+
+.. NOTE: You can use the :pr:`xx` and :issue:`xx` role to x-ref to a GitHub PR
+   or issue from this project.
+
+:hide-toc:
+
+.. include:: ./authors.inc
+
+Version 1.3
+===========
+
+- Improve logging to set MNE's logger level to currently set MNE-LSL level in MNE-LSL functions using MNE's function under the hood (:pr:`219` by `Mathieu Scheltienne`_)
+- Use the currently set logger level as replacement for ``verbose=None`` in MNE-LSL functions (:pr:`219` by `Mathieu Scheltienne`_)
+- Add IIR filter support to the LSL stream through :meth:`mne_lsl.stream.StreamLSL.filter` (:pr:`218` by `Mathieu Scheltienne`_ and `Eric Larson`_)
+- Add IIR notch filter support to the LSL stream through :meth:`mne_lsl.stream.StreamLSL.notch_filter` (:pr:`226` by `Mathieu Scheltienne`_)
+
+API changes (backward incompatible)
+-----------------------------------
+
+- Add ``n_repeat`` argument to :class:`~mne_lsl.player.PlayerLSL` to support looping a file a fix number of times (:pr:`223` by `Mathieu Scheltienne`_)
+
+.. _latest:
+
+Version 1.3.1
+=============
+
+- Fix handling of warnings through the logger (:pr:`243` by `Mathieu Scheltienne`_)
