+++ stable/_sources/generated/tutorials/30_stream_manual.rst.txt
@@ -1,0 +1,629 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "generated/tutorials/30_stream_manual.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_generated_tutorials_30_stream_manual.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_generated_tutorials_30_stream_manual.py:
+
+
+Automatic vs Manual acquisition
+===============================
+
+.. include:: ./../../links.inc
+
+The :class:`~mne_lsl.stream.StreamLSL` object offers 2 mode of acquisition: automatic or
+manual. In automatic mode, the stream object acquires new chunks of samples at a
+regular interval. In manual mode, the user has to call the
+:meth:`~mne_lsl.stream.StreamLSL.acquire` to acquire new chunks of samples from the
+network. The automatic or manual acquisition is selected via the ``acquisition_delay``
+argument of :meth:`~mne_lsl.stream.StreamLSL.connect`:
+
+- a non-zero positive integer value will set the acquisition to automatic mode with the
+  specified delay in seconds.
+- ``0`` will set the acquisition to manual mode.
+
+Automatic acquisition
+---------------------
+
+When the stream is set to automatically acquire new samples at a regular interval, a
+background thread is created with :class:`concurrent.futures.ThreadPoolExecutor`. The
+background thread is periodically receives a job to acquire new samples from the
+network.
+
+.. important::
+
+    If the main thread is hogging all of the CPU resources, the delay between two
+    acquisition job might be longer than the specified delay. The background thread
+    will always do its best to acquire new samples at the specified delay, but it is not
+    able to do so if the CPU is busy.
+
+.. GENERATED FROM PYTHON SOURCE LINES 33-49
+
+.. code-block:: Python
+
+
+    import uuid
+    from time import sleep
+
+    from matplotlib import pyplot as plt
+
+    from mne_lsl.datasets import sample
+    from mne_lsl.player import PlayerLSL
+    from mne_lsl.stream import StreamLSL
+
+    # create a mock LSL stream for this tutorial
+    fname = sample.data_path() / "sample-ant-raw.fif"
+    source_id = uuid.uuid4().hex
+    player = PlayerLSL(fname, chunk_size=200, source_id=source_id).start()
+    player.info
+
+
+
+
+
+
+.. raw:: html
+
+    <div class="output_subarea output_html rendered_html output_result">
+    <script type="text/javascript">
+        const toggleVisibility = (className) => {
+
+      const elements = document.querySelectorAll(`.${className}`)
+
+      elements.forEach(element => {
+        if (element.classList.contains('repr-section-header')) {
+          // Don't collapse the section header row.
+           return
+        }
+        if (element.classList.contains('repr-element-collapsed')) {
+          // Force a reflow to ensure the display change takes effect before removing the class
+          element.classList.remove('repr-element-collapsed')
[…]
