### 2024-12-23 224dfae
Bump astral-sh/setup-uv from 4 to 5 (#384)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
