+++ .github/workflows/stubs.yaml
@@ -19,7 +19,7 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: 3.13
-      - uses: astral-sh/setup-uv@v4
+      - uses: astral-sh/setup-uv@v5
       - run: uv pip install -e .[stubs] --quiet --system
       - run: mne-lsl sys-info --developer
       - run: python tools/stubgen.py
