### 2025-01-13 5b1fd6d
[pre-commit.ci] pre-commit autoupdate (#387)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
