+++ src/mne_lsl/utils/config.py
@@ -41,9 +41,9 @@ def sys_info(fid: IO | None = None, developer: bool = False):
     out("Logical cores:".ljust(ljust) + str(psutil.cpu_count(True)) + "\n")
     # memory information
     out("RAM:".ljust(ljust))
-    out(f"{psutil.virtual_memory().total / float(2 ** 30):0.1f} GB\n")
+    out(f"{psutil.virtual_memory().total / float(2**30):0.1f} GB\n")
     out("SWAP:".ljust(ljust))
-    out(f"{psutil.swap_memory().total / float(2 ** 30):0.1f} GB\n")
+    out(f"{psutil.swap_memory().total / float(2**30):0.1f} GB\n")
     # package information
     out(f"{package}:".ljust(ljust) + version(package) + "\n")
     try:
