+++ tutorials/00_introduction.py
@@ -71,7 +71,7 @@ Mocking an LSL stream
 
 To build real-time applications or showcase their functionalities, such as in this
 tutorial, it's essential to generate simulated LSL streams. This involves creating a
-:class:~mne_lsl.lsl.StreamOutlet and regularly sending data through it.
+:class:`~mne_lsl.lsl.StreamOutlet` and regularly sending data through it.
 
 Within MNE-LSL, the :class:`~mne_lsl.player.PlayerLSL` generates a simulated LSL stream
 utilizing data from a :class:`mne.io.Raw` file or object. This stream inherits its
