### 2025-01-18 c26bde8
add changelog entry
