### 2025-01-18 93eefa9
Fix intersphinx link in tutorial (#389)
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
