### 2025-02-21 d275efa
Fix for annotation with no duration (#399)
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
