+++ tutorials/20_player_annotations.py
@@ -22,6 +22,11 @@ description: ``'event1'``, ``'event2'``, and ``'event3'``. The event stream will
 channels, each corresponding to one of the 3 descriptions. When an annotation is
 streamed, it's duration is encoded as the value on its channel while the other channels
 remain to zero.
+
+.. note::
+
+    Annotation with a duration equal to zero are special cased and yield an encoded
+    value of ``-1``.
 """
 
 # sphinx_gallery_thumbnail_number = 2
