### 2025-02-21 #7228 @sappelhoff
@Kevin-Mattheus-Moerman unfortunately I have a conflict of interest for this one, because @mscheltienne and I are both on the maintainers team for mne-python. However, if you want to wave my conflict of interest, I would be happy to serve as an editor.
