+++ doc/development/changes/latest.rst
@@ -14,9 +14,9 @@
 
 .. include:: ./authors.inc
 
-Version 1.9
-===========
+.. _latest:
 
-- Fix sphinx formatting in tutorials (:pr:`388` and :pr:`389` by `Thomas Binns`_)
-- Fix source links in the documentation (:pr:`398` by `Tom Ma`_)
-- Fix stream of annotations by a :class:`~mne_lsl.player.PlayerLSL` when the annotation duration is 0 (:pr:`399` by `Mathieu Scheltienne`_)
+Version 1.10
+============
+
+- xxx
