+++ 1.0.0/_sources/generated/tutorials/20_stream_meas_info.rst.txt
@@ -1,0 +1,365 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "generated/tutorials/20_stream_meas_info.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_generated_tutorials_20_stream_meas_info.py>`
+        to download the full example code
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_generated_tutorials_20_stream_meas_info.py:
+
+
+Stream information
+==================
+
+.. include:: ./../../links.inc
+
+A :class:`~mne_lsl.stream.StreamLSL` will automatically attempt to interpret the channel
+names, types and units during the connection with
+:meth:`mne_lsl.stream.StreamLSL.connect`.
+However, by definition, an LSL stream does not require any of those information to be
+present. Moreover, the channel type and unit are not standardize, and may be define with
+different nomenclature depending on the system and the application emitting the LSL
+stream. For instance, an EEG channel might be denoted by the type ``'eeg'`` or
+``'electroencephalography'``, or something else entirely.
+
+If ``MNE-LSL`` is not able to interpret the information in a stream description, it will
+default to:
+
+* numbers instead of channel names if it failed to load the channel names, ``'0'```,
+  ``'1'``, ... as :func:`mne.create_info` does when the argument ``ch_names`` is
+  provided as a number of channels.
+* The stream type (if interpretable) or ``'misc'`` otherwise if it failed to load the
+  individual channel types.
+* SI units (factor 0) if it failed to load the individual channel units.
+
+The stream and channel type supported correspond to the MNE-supported channel types.
+
+.. GENERATED FROM PYTHON SOURCE LINES 30-46
+
+Inspecting a stream info
+------------------------
+
+A :class:`~mne_lsl.stream.StreamLSL` measurement information can be inspected with
+similar methods to a :class:`~mne.io.Raw` object:
+:py:attr:`mne_lsl.stream.StreamLSL.info`,
+:py:attr:`mne_lsl.stream.StreamLSL.ch_names`,
+:meth:`mne_lsl.stream.StreamLSL.get_channel_types`,
+:meth:`mne_lsl.stream.StreamLSL.get_channel_units`.
+
+.. note::
+
+    For this tutorial purposes, a mock LSL stream is created using a
+    :class:`~mne_lsl.player.PlayerLSL`. See
+    :ref:`sphx_glr_generated_tutorials_10_player.py` for additional information on
+    mock LSL streams.
+
+.. GENERATED FROM PYTHON SOURCE LINES 46-58
+
+.. code-block:: default
+
+
+    from mne_lsl.datasets import sample
+    from mne_lsl.player import PlayerLSL as Player
+    from mne_lsl.stream import StreamLSL as Stream
+
+    fname = sample.data_path() / "sample-ant-aux-raw.fif"
+    player = Player(fname)
+    player.start()
+    stream = Stream(bufsize=5)  # 5 seconds of buffer
+    stream.connect()
+    stream.info
+
+
+
+
+
+
+.. raw:: html
+
+    <div class="output_subarea output_html rendered_html output_result">
+    <table class="table table-hover table-striped table-sm table-responsive small">
+        <tr>
[…]
