### 2025-03-17 19cd430
Bump pypa/cibuildwheel from 2.23.0 to 2.23.1 (#406)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
