### 2025-03-31 e3e03b9
Bump pypa/cibuildwheel from 2.23.1 to 2.23.2 (#410)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
