### 2025-04-22 #7228 @mstimberg
@matthiasdold Apologies, but as the friendly bot said, you can only generate the checklist in the actual review, and we are still in the *pre*-review since I haven't found a second reviewer yet. But, given that things are taking a lot longer than everybody would like, I am going to start the review process now so that you can already start working on your review while I continue my search for a second reviewer. Thanks to everyone for your patience 🙏 !
