### 2025-04-22 #8088 @editorialbot
[REVIEW]: MNE-LSL: Real-time framework integrated with MNE-Python for online neuroscience research through LSL-compatible devices.
**Submitting author:** @mscheltienne (<a href="http://orcid.org/0000-0001-8316-7436">Mathieu Scheltienne</a>)
**Repository:** https://github.com/mne-tools/mne-lsl
**Branch with paper.md** (empty if default branch):
**Version:** 1.5.0
