### 2025-04-22 #8088 @mstimberg
👋🏼  @mscheltienne @matthiasdold this is the review thread for the paper. All of our communications will happen here from now on.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/8088#issue-3011872647).
