### 2025-04-23 df5f9ac
improve docstring
