+++ src/mne_lsl/stream/epochs.py
@@ -113,7 +113,8 @@ class EpochsStream:
       ignored. This last case can be useful when working with a ``Player`` replaying
       annotations from a file as one-hot encoded events.
 
-    Event streams irregularly sampled and a ``str`` datatype are not yet supported.
+    Event streams irregularly sampled and using a ``str`` datatype are not yet
+    supported.
 
     .. note::
 
