### 2025-04-28 3bfc5a3
Bump pypa/cibuildwheel from 2.23.2 to 2.23.3 (#417)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@dandelion.science>
