+++ tests/stream/test_epochs.py
@@ -30,11 +30,12 @@ from mne_lsl.stream.epochs import (
 if TYPE_CHECKING:
     from collections.abc import Generator
 
+    from mne import Info
     from mne.io import BaseRaw
     from numpy.typing import NDArray
 
 
-def test_ensure_event_id():
+def test_ensure_event_id() -> None:
     """Test validation of event dictionary."""
     assert _ensure_event_id(5, None) == {"5": 5}
     assert _ensure_event_id({"5": 5}, None) == {"5": 5}
@@ -54,7 +55,7 @@ def test_ensure_event_id():
         _ensure_event_id({"101": None}, None)
 
 
-def test_check_baseline():
+def test_check_baseline() -> None:
     """Test validation of the baseline."""
     _check_baseline(None, -0.2, 0.5)
     _check_baseline((None, None), -0.2, 0.5)
@@ -78,14 +79,14 @@ def test_check_baseline():
 
 
 @pytest.fixture(scope="session")
-def info():
+def info() -> Info:
     """A simple info object with 10 EEG channels."""
     return create_info(
         ch_names=[f"ch{i}" for i in range(10)], sfreq=100, ch_types="eeg"
     )
 
 
-def test_check_reject_flat(info):
+def test_check_reject_flat(info: Info) -> None:
     """Test validation of the rejection dictionaries."""
     _check_reject_flat(None, None, info)
     _check_reject_flat(None, {"eeg": 1e-6}, info)
@@ -111,7 +112,7 @@ def test_check_reject_flat(info):
         _check_reject_flat(None, {"eeg": "test"}, info)
 
 
-def test_check_reject_tmin_tmax():
+def test_check_reject_tmin_tmax() -> None:
     """Test validation of rejection time windows."""
     _check_reject_tmin_tmax(None, None, -0.2, 0.5)
     _check_reject_tmin_tmax(-0.2, 0.5, -0.2, 0.5)
@@ -132,7 +133,7 @@ def test_check_reject_tmin_tmax():
         _check_reject_tmin_tmax(None, "test", -0.2, 0.5)
 
 
-def test_ensure_detrend_str():
+def test_ensure_detrend_str() -> None:
     """Test validation of detrend."""
     assert _ensure_detrend_str(None) is None
     assert _ensure_detrend_str(0) == "constant"
@@ -204,7 +205,7 @@ def events() -> NDArray[np.int64]:
     return events
 
 
-def test_prune_events(events: NDArray[np.int64]):
+def test_prune_events(events: NDArray[np.int64]) -> None:
     """Test pruning events."""
     ts = np.arange(10000, 11000, 1.8)
     events_ = _prune_events(events, dict(a=1, b=2, c=3), 10, ts, None, None, 0)
@@ -271,7 +272,7 @@ def raw_with_stim_channel() -> BaseRaw:
 class DummyPlayer:
     """Dummy player object containing the player attributes."""
 
-    def __init__(self, /, **kwargs):
+    def __init__(self, /, **kwargs) -> None:
         self.__dict__.update(kwargs)
 
 
@@ -295,7 +296,9 @@ def _player_mock_lsl_stream(
 
 
 @pytest.fixture
-def mock_lsl_stream(raw_with_stim_channel, request, chunk_size):
+def mock_lsl_stream(
+    raw_with_stim_channel: BaseRaw, request: pytest.FixtureRequest, chunk_size: int
+) -> Generator[DummyPlayer, None, None]:
     """Create a mock LSL stream streaming events on a stim channel."""
     manager = mp.Manager()
     status = manager.Value("i", 0)
@@ -314,7 +317,7 @@ def mock_lsl_stream(raw_with_stim_channel, request, chunk_size):
     process.kill()
 
 
[…]
