+++ tests/utils/test_time.py
@@ -1,9 +1,11 @@
+from __future__ import annotations
+
 import time
 
 from mne_lsl.utils._time import high_precision_sleep
 
 
-def test_high_precision_sleep():
+def test_high_precision_sleep() -> None:
     """Test high precision sleep function."""
     start = time.perf_counter()
     high_precision_sleep(0.4)
