+++ tests/datasets/test_datasets.py
@@ -12,11 +12,12 @@ from mne_lsl.utils._tests import sha256sum
 
 if TYPE_CHECKING:
     from pathlib import Path
+    from types import ModuleType
 
 
 @pytest.mark.xfail(reason="Connection issue to the dataset servers.")
 @pytest.mark.parametrize("dataset", [sample, testing])
-def test_data_path(dataset) -> None:
+def test_data_path(dataset: ModuleType) -> None:
     """Test download if the testing dataset."""
     if dataset != testing and os.getenv("GITHUB_ACTIONS", "") == "true":
         pytest.skip("Skip sample dataset download on GitHub Actions.")
@@ -28,7 +29,7 @@ def test_data_path(dataset) -> None:
 @pytest.mark.parametrize(
     ("dataset", "registry"), [(sample, _REGISTRY_SAMPLE), (testing, _REGISTRY_TESTING)]
 )
-def test_make_registry(tmp_path: Path, dataset, registry: Path) -> None:
+def test_make_registry(tmp_path: Path, dataset: ModuleType, registry: Path) -> None:
     """Test the registry tmp_path making."""
     if dataset != testing and os.getenv("GITHUB_ACTIONS", "") == "true":
         pytest.skip("Skip sample dataset download on GitHub Actions.")
