+++ tests/lsl/test_stream_inlet.py
@@ -68,7 +68,9 @@ def _test_numerical_data(
         ("int32", np.int32),
     ],
 )
-def test_pull_numerical_sample(dtype_str: str, dtype, close_io: Callable) -> None:
+def test_pull_numerical_sample(
+    dtype_str: str, dtype: np.dtype, close_io: Callable
+) -> None:
     """Test pull_sample with numerical values."""
     x = np.array([1, 2], dtype=dtype)
     assert x.shape == (2,)
@@ -123,7 +125,9 @@ def test_pull_str_sample(close_io: Callable) -> None:
         ("int32", np.int32),
     ],
 )
-def test_pull_numerical_chunk(dtype_str: str, dtype, close_io: Callable) -> None:
+def test_pull_numerical_chunk(
+    dtype_str: str, dtype: np.dtype, close_io: Callable
+) -> None:
     """Test pull_chunk on a numerical chunk."""
     x = np.array([[1, 4], [2, 5], [3, 6]], dtype=dtype)
     assert x.shape == (3, 2)
@@ -230,7 +234,7 @@ def test_get_sinfo(close_io: Callable) -> None:
         ("int32", np.int32),
     ],
 )
-def test_inlet_methods(dtype_str: str, dtype, close_io: Callable) -> None:
+def test_inlet_methods(dtype_str: str, dtype: np.dtype, close_io: Callable) -> None:
     """Test the methods from an Inlet."""
     x = np.array([[1, 4], [2, 5], [3, 6]], dtype=dtype)
     assert x.shape == (3, 2)
