### 2025-05-05 31e3cbb
Improve type-hints through unit tests (#421)
