+++ tests/datasets/test_datasets.py
@@ -1,4 +1,7 @@
+from __future__ import annotations
+
 import os
+from typing import TYPE_CHECKING
 
 import pytest
 
@@ -7,10 +10,14 @@ from mne_lsl.datasets.sample import _REGISTRY as _REGISTRY_SAMPLE
 from mne_lsl.datasets.testing import _REGISTRY as _REGISTRY_TESTING
 from mne_lsl.utils._tests import sha256sum
 
+if TYPE_CHECKING:
+    from pathlib import Path
+    from types import ModuleType
+
 
 @pytest.mark.xfail(reason="Connection issue to the dataset servers.")
 @pytest.mark.parametrize("dataset", [sample, testing])
-def test_data_path(dataset):
+def test_data_path(dataset: ModuleType) -> None:
     """Test download if the testing dataset."""
     if dataset != testing and os.getenv("GITHUB_ACTIONS", "") == "true":
         pytest.skip("Skip sample dataset download on GitHub Actions.")
@@ -22,7 +29,7 @@ def test_data_path(dataset):
 @pytest.mark.parametrize(
     ("dataset", "registry"), [(sample, _REGISTRY_SAMPLE), (testing, _REGISTRY_TESTING)]
 )
-def test_make_registry(tmp_path, dataset, registry):
+def test_make_registry(tmp_path: Path, dataset: ModuleType, registry: Path) -> None:
     """Test the registry tmp_path making."""
     if dataset != testing and os.getenv("GITHUB_ACTIONS", "") == "true":
         pytest.skip("Skip sample dataset download on GitHub Actions.")

[…]
