### 2025-05-05 a4a0668
improve close_io type hint
