### 2025-05-05 c00ca5b
add setuptools to dependencies
