### 2025-05-06 #8088 @mstimberg
Thanks again for helping out!

As mentioned in my [comment above](https://github.com/openjournals/joss-reviews/issues/8088#issuecomment-2822207994), the first step is is to create a checklist for your review by entering
at the start of a new comment in this thread. More info and links in the [top-most comment](https://github.com/openjournals/joss-reviews/issues/8088#issue-3011872647). And of course, feel free to ping me if there are any questions.

@matthiasdold don't hesitate to create your checklist either now that we are in the review issue 😊
