+++ pyproject.toml
@@ -163,7 +163,7 @@ junit_family = 'xunit2'
 minversion = '8.0'
 
 [tool.ruff]
-extend-exclude = []
+extend-exclude = ["doc/_examples/*.py"]
 line-length = 88
 
 [tool.ruff.format]
