### 2025-05-06 f45e551
Add support for callback(s) (#419)
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
