### 2025-05-06 f5f4e1e
fix docstring typo
