+++ src/mne_lsl/lsl/functions.py
@@ -78,7 +78,7 @@ def resolve_streams(
     source_id : str | None
         Restrict the selected stream to this source ID.
     minimum : int
-        Minimum number of stream to return where restricting the selection. As soon as
+        Minimum number of stream to return when restricting the selection. As soon as
         this minimum is hit, the search will end. Only works if at least one of the 3
         identifiers ``name``, ``stype`` or ``source_id`` is not ``None``.
 
