### 2025-05-07 3464863
add support to hash streams
