### 2025-05-07 ef0ad3c
Add support for stream hashing and usage as dictionary keys (#424)
