+++ 1.7.1/_sources/generated/examples/index.rst.txt
@@ -1,0 +1,137 @@
+:orphan:
+
+.. include:: ../../links.inc
+
+Examples
+========
+
+This page contains examples using ``MNE-LSL``.
+
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbnails">
+
+.. thumbnail-parent-div-open
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="With a PlayerLSL object, you can replay an MNE-Python readable file as a valid LSL stream. This is very useful for testing or debugging purposes. The created LSL stream is managed in the main process, as a separate thread. Thus, if the code running the main process is hogging all of the CPU resources, then the PlayerLSL object may not be able to keep up with the LSL stream and to push (publish) new data when it should. This is why it is often useful to run the PlayerLSL object in a separate process.">
+
+.. only:: html
+
+  .. image:: /generated/examples/images/thumb/sphx_glr_00_player_separate_process_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_generated_examples_00_player_separate_process.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Replaying a file in a separate process</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="With a StreamLSL, we can build a real-time peak detector. The structure defined below will be adapted for cardiac R-peak detection, but remains valid for other peak detector. Note however that the peak detection performance, i.e. how fast it will be able to detect a peak, will depend heavily on the peak shape.">
+
+.. only:: html
+
+  .. image:: /generated/examples/images/thumb/sphx_glr_10_peak_detection_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_generated_examples_10_peak_detection.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Real-time peak detection</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="With a StreamLSL, we can compute the bandpower on a time rolling window. For this example, we will look at the alpha band power, between 8 and 13 Hz.">
+
+.. only:: html
+
+  .. image:: /generated/examples/images/thumb/sphx_glr_20_bandpower_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_generated_examples_20_bandpower.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Bandpower rolling window</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="With a EpochsStream, we can build a real-time evoked response visualization. This is useful to monitor the brain activity in real-time.">
+
+.. only:: html
+
+  .. image:: /generated/examples/images/thumb/sphx_glr_30_real_time_evoked_responses_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_generated_examples_30_real_time_evoked_responses.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Real-time evoked responses</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="This example demonstrates how to decode real-time data using MNE-Python and Scikit-learn. We will stream the sample_audvis_raw.fif file from MNE&#x27;s sample dataset with a PlayerLSL, process the signal through a StreamLSL, and decode the epochs created with EpochsStream.">
+
+.. only:: html
+
[…]

[…]
