### 2025-05-12 1dfc70a
test fix for click>=8.2.0
