+++ 1.6.1/_sources/generated/api/mne_lsl.player.PlayerLSL.rst.txt
@@ -1,0 +1,11 @@
+﻿mne\_lsl.player.PlayerLSL
+=========================
+
+.. currentmodule:: mne_lsl.player
+
+.. autoclass:: PlayerLSL
+   :members:
+   :inherited-members:
+
+.. minigallery:: mne_lsl.player.PlayerLSL
+    :add-heading:
