+++ 1.4.0/_sources/api/utilities.rst.txt
@@ -1,0 +1,23 @@
+Utilities
+---------
+
+Logging utilities are available to interact with ``mne_lsl`` logger.
+
+.. currentmodule:: mne_lsl
+
+.. autosummary::
+    :toctree: ../generated/api
+    :nosignatures:
+
+    add_file_handler
+    set_log_level
+
+Development utilities are available to help debug a setup.
+
+.. currentmodule:: mne_lsl
+
+.. autosummary::
+    :toctree: ../generated/api
+    :nosignatures:
+
+    sys_info

[…]
