### 2025-05-12 d181499
[pre-commit.ci] pre-commit autoupdate (#425)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
