### 2025-05-18 mne-tools/mne-lsl#426 @matthiasdold
Code for reproducing performance benchmarks
Hi mne-lsl team,

I would like to reproduce the [benchmark measurements](https://mne.tools/mne-lsl/stable/resources/pylsl.html#differences-with-pylsl) you did for the comparison of `mne-lsl` vs `pylsl`, as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8088). Maybe I have missed it, but I do not find a code example for running the benchmark.

I have seen [this timeit comment](https://github.com/mne-tools/mne-lsl/blob/d18149921e61f43c1014d4e5e4821e44bcd2944c/src/mne_lsl/lsl/stream_inlet.py#L400) but as this concerns the transformation from the buffer to either list or NDarray, I am not sure if this is what you were referring to in the [low-level API references](https://mne.tools/mne-lsl/stable/api/lsl%20%28low-level%29.html). Both your implementation and `pylsl` are calling the same liblsl function according to the [`fmt2pull_chunk`](https://github.com/mne-tools/mne-lsl/blob/d18149921e61f43c1014d4e5e4821e44bcd2944c/src/mne_lsl/lsl/constants.py#L110) map.

So comparing `mne-lsl` with `pylsl` with pre-allocated buffer, to avoid the slow list comprehension, I would see the differences in the two approaches:

**mne-lsl**'s approach of: [creating the buffer](https://github.com/mne-tools/mne-lsl/blob/d18149921e61f43c1014d4e5e4821e44bcd2944c/src/mne_lsl/lsl/stream_inlet.py#L127) as c_array + pulling chunk to buffer +[np.frombuffer to `sample` NDarray](https://github.com/mne-tools/mne-lsl/blob/d18149921e61f43c1014d4e5e4821e44bcd2944c/src/mne_lsl/lsl/stream_inlet.py#L396).

**pylsl**'s approach of: [creating a buffer and linking it to a `dest_object` (i.e., NDarray)](https://github.com/labstreaminglayer/pylsl/blob/5e88eac4a4f82a809e0560fd4623e1735b3f57bf/src/pylsl/inlet.py#L247) + pulling chunk to buffer.

Basically, it is about `np.frombuffer` during each pull, vs declaring the buffer as `(self.value_type * max_values).from_buffer(dest_obj)` up front once....

This is just focusing on numeric types, as string/chars will be slow regardless.

It would also be great to see the speed dependency with regards to samples (here understood as `n_channels` x `n_samples_per_channel`) per second. The pre-allocated approach becomes essential when number of samples  per pull become large. I remember from my own benchmarking, that the list comprehension approach even breaks down rather soon, while the pre-allocated could process ~500 channels at about 22kHz.
