### 2025-05-19 mne-tools/mne-lsl#426 @mscheltienne
Indeed, `mne-lsl` and `pylsl` use the same underlying `liblsl` function. There is no performance differencec at this stage, but there are 3 ways (I know of) to retrieve the buffer in Python:

- list comprehension: slow but also the default behavior of `pylsl`, and ASFAIK the most commonly used because many people miss the `dest_obj` argument.
- using a destination object (argument `dest_obj` in `pylsl`)
- using `numpy.frombuffer`, default behavior of `mne-lsl`

It's also true that this improvement is only present for numerical types, as mentioned in the note at the bottom of [this section](Note that this performance improvement is absent for string based streams. Follow [#225](https://github.com/mne-tools/mne-lsl/issues/225) for more information.):

> Note that this performance improvement is absent for string based streams. Follow [#225](https://github.com/mne-tools/mne-lsl/issues/225) for more information.

I'll add a code snippet to reproduce, IIRC it's in one of the PRs.. and I'll also highlight the note above. The speed dependency is mentioned as the list comprehension runs in `O(n)` while `np.frombuffer` runs in `O(1)` (from empirical measurements). I'll add those to the code snippet to reproduce.
