+++ doc/resources/pylsl.rst
@@ -9,9 +9,10 @@ Faster chunk pull
 -----------------
 
 Arguably the most important difference, pulling a chunk of numerical data with
-:meth:`~mne_lsl.lsl.StreamInlet.pull_chunk` is much faster than with its
-`pylsl <lsl python_>`_ counterpart. By default, `pylsl <lsl python_>`_ loads the
-retrieved samples one by one in a list of list, `here <pylsl pull_chunk_>`_.
+:meth:`~mne_lsl.lsl.StreamInlet.pull_chunk` is faster than with its
+`pylsl <lsl python_>`_ counterpart, except if using the argument ``dest_obj``.
+By default, `pylsl <lsl python_>`_ loads the retrieved samples one by one in a list of
+list, `here <pylsl pull_chunk_>`_.
 
 .. code-block:: python
 
@@ -41,8 +42,74 @@ Now, ``samples`` is created in constant time ``O(1)``. The performance gain vari
 depending on the number of values pulled, for instance retrieving 1024 samples with
 65 channels in double precision (``float64``) takes:
 
-* 4.33 ms ± 37.5 µs with ``pylsl`` (default behavior)
-* 268 ns ± 0.357 ns with ``mne_lsl.lsl``
+* 3.91 ± 0.12 ms with ``pylsl`` (default behavior: list comprehension)
+* 419.08 ± 38.10 ns with ``mne_lsl.lsl`` (:func:`numpy.frombuffer`)
+
+Increasing the number of channels to 650, simulating an higher sample count yields:
+
+* 46.56 ± 1.06 ms per loop with ``pylsl`` (default behavior: list comprehension)
+* 424.29 ± 36.41 ns with ``mne_lsl.lsl`` (:func:`numpy.frombuffer`)
+
+.. dropdown:: Timeit python code
+    :animate: fade-in-slide-down
+
+    .. coode-block:: python
+
+        import ctypes
+        import timeit
+
+        import numpy as np
+
+        n_samples = 1024
+        n_channels = 650
+        max_values = n_samples * n_channels
+
+        # create a data buffer, 'retrieved from liblsl'
+        data_buffer = (ctypes.c_double * max_values)()
+        for i in range(max_values):
+            data_buffer[i] = float(i)
+
+
+        def method_list_comprehension():
+            """List comprehension, default method for pylsl."""
+            samples = [
+                [data_buffer[s * n_channels + c] for c in range(n_channels)]
+                for s in range(n_samples)
+            ]
+            return samples
+
+
+        def method_frombuffer():
+            """Numpy frombuffer, default method for mne-lsl."""
+            samples = np.frombuffer(data_buffer, dtype=np.float64)[:max_values].reshape(
+                n_samples, n_channels
+            )
+            return samples
+
+
+        repeat = 5
+        number = 100
+        for func in (method_list_comprehension, method_frombuffer):
+            timer = timeit.Timer(func)
+            results = timer.repeat(repeat, number=number)
+            times = [t / number for t in results]
+            best = min(times)
+            worst = max(times)
+            avg = sum(times) / len(times)
+
+            # format with the corret unit
+            if best < 1e-6:
+                unit, factor = "ns", 1e9
+            elif best < 1e-3:
+                unit, factor = "µs", 1e6
+            elif best < 1:
+                unit, factor = "ms", 1e3
+            else:
+                unit, factor = "s", 1
+
+            formatted_time = f"{best * factor:.2f} ± {(avg - best) * factor:.2f} {unit}"
+            plural = "s" if number != 1 else ""
+            print(f"{number} loop{plural}, best of {repeat}: {formatted_time} per loop")
 
[…]
