+++ 1.4.0/_sources/generated/tutorials/20_player_annotations.rst.txt
@@ -1,0 +1,55 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "generated/tutorials/20_player_annotations.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_generated_tutorials_20_player_annotations.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_generated_tutorials_20_player_annotations.py:
+
+
+.. _tut-player-annotations:
+
+Player with annotations
+=======================
+
+.. include:: ./../../links.inc
+
+
+.. rst-class:: sphx-glr-timing
+
+   **Total running time of the script:** (0 minutes 0.000 seconds)
+
+**Estimated memory usage:**  9 MB
+
+
+.. _sphx_glr_download_generated_tutorials_20_player_annotations.py:
+
+.. only:: html
+
+  .. container:: sphx-glr-footer sphx-glr-footer-example
+
+    .. container:: sphx-glr-download sphx-glr-download-jupyter
+
+      :download:`Download Jupyter notebook: 20_player_annotations.ipynb <20_player_annotations.ipynb>`
+
+    .. container:: sphx-glr-download sphx-glr-download-python
+
+      :download:`Download Python source code: 20_player_annotations.py <20_player_annotations.py>`
+
+
+.. only:: html
+
+ .. rst-class:: sphx-glr-signature
+
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
