+++ 1.4.0/_sources/changes/index.rst.txt
@@ -1,0 +1,11 @@
+Changelog
+=========
+
+.. toctree::
+    :titlesonly:
+
+    1.4.rst
+    1.3.rst
+    1.2.rst
+    1.1.rst
+    1.0.rst
