### 2025-05-22 mne-tools/mne-lsl#426 @mscheltienne
Following the PR, the documentation has been updated at https://mne.tools/mne-lsl/dev/resources/pylsl.html#differences-with-pylsl (will be moved to the default `stable` during the next release process), including a python code snippet to reproduce the timing measurements.
