### 2025-05-27 mne-tools/mne-lsl#430 @teonbrooks
Update citation for LSL package to use Github repository link
Hi there mne-lsl team,

The issue is raised in review of https://github.com/openjournals/joss-reviews/issues/8088

According to LSL's FAQ, the preferred citation for the project is their Github repository:
https://labstreaminglayer.readthedocs.io/info/faqs.html#referencing-lsl

In the paper, it currently references the project website.
