### 2025-05-29 5a6cc22
ADD keyword, FIX install command
