+++ pyproject.toml
@@ -38,6 +38,7 @@ keywords = [
   'EEG',
   'eeg',
   'electroencephalography',
+  'neurophysiology',
   'labstreaminglayer',
   'LSL',
   'neuroimaging',
