+++ pyproject.toml
@@ -38,10 +38,10 @@ keywords = [
   'EEG',
   'eeg',
   'electroencephalography',
-  'neurophysiology',
   'labstreaminglayer',
   'LSL',
   'neuroimaging',
+  'neurophysiology',
   'neuroscience',
   'python',
   'real-time',
