### 2025-05-30 a3f803f
update licence file format
