+++ pyproject.toml
@@ -46,8 +46,7 @@ keywords = [
   'python',
   'real-time',
 ]
-license = 'BSD-3-Clause'
-license-files = ['LICENSE']
+license = {file = 'LICENSE'}
 maintainers = [
   {email = 'mathieu.scheltienne@fcbg.ch', name = 'Mathieu Scheltienne'},
 ]
