### 2025-05-30 4958296
Improve the contributing guidelines to make an editable install of mne-lsl (#433)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@dandelion.science>
