+++ 1.6.0/_sources/generated/tutorials/40_epochs.rst.txt
@@ -1,0 +1,3038 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "generated/tutorials/40_epochs.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_generated_tutorials_40_epochs.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_generated_tutorials_40_epochs.py:
+
+
+Epoching a Stream in real-time
+==============================
+
+.. include:: ./../../links.inc
+
+The :class:`~mne_lsl.stream.EpochsStream` object can be used similarly to
+:class:`mne.Epochs` to create epochs from a continuous stream of samples around events
+of interest.
+
+.. note::
+
+    The :class:`~mne_lsl.stream.EpochsStream` object is designed to work with
+    any ``Stream`` object. At the time of writing, only
+    :class:`~mne_lsl.stream.StreamLSL` is available, but any object inheriting from the
+    abstract :class:`~mne_lsl.stream.BaseStream` object should work.
+
+A :class:`~mne_lsl.stream.EpochsStream` object support peak-to-peak rejection, baseline
+correction and detrending.
+
+Event source
+------------
+
+A :class:`~mne_lsl.stream.EpochsStream` object requires an event source to create
+epochs. 3 event sources are supported:
+
+- a set of ``'stim'`` channels within the attached ``Stream`` object.
+- a set of ``'stim'`` channels within a separate ``Stream`` object.
+- an irregularly sampled ``Stream`` object.
+
+.. note::
+
+    In the case of an irregularly sampled ``Stream`` object, only numerical streams are
+    supported at the moment because interaction with ``str`` streams in Python is not
+    as performant as interaction with numerical streams.
+
+Set of ``'stim'`` channels
+--------------------------
+
+The set of ``'stim'`` channels from which the events are extracted can be either part
+of the regularly sampled ``Stream`` object epoched (argument ``stream``) or part of a
+separate regularly sampled ``Stream`` object (argument ``event_stream``). In both case,
+the channel(s) type should be ``'stim'`` and the channel(s) should be formatted for
+:func:`mne.find_events` to correctly extract the events. The channels to consider are
+provided in the argument ``event_channels`` and the events to consider in the argument
+``event_id``. Let's create epochs around the event ID ``2`` from the ``'STI 014'``
+channel of MNE's sample dataset.
+
+.. GENERATED FROM PYTHON SOURCE LINES 49-75
+
+.. code-block:: Python
+
+
+    import time
+    import uuid
+
+    import numpy as np
+    from matplotlib import pyplot as plt
+    from mne import Epochs, EpochsArray, annotations_from_events, find_events
+    from mne.io import read_raw_fif
+
[…]
