+++ stable/_sources/generated/examples/00_player_separate_process.rst.txt
@@ -1,0 +1,280 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "generated/examples/00_player_separate_process.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_generated_examples_00_player_separate_process.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_generated_examples_00_player_separate_process.py:
+
+
+Replaying a file in a separate process
+======================================
+
+With a :class:`~mne_lsl.player.PlayerLSL` object, you can replay an MNE-Python readable
+file as a valid LSL stream. This is very useful for testing or debugging purposes. The
+created LSL stream is managed in the main process, as a separate thread. Thus, if the
+code running the main process is hogging all of the CPU resources, then the
+:class:`~mne_lsl.player.PlayerLSL` object may not be able to keep up with the LSL stream
+and to push (publish) new data when it should. This is why it is often useful to run the
+:class:`~mne_lsl.player.PlayerLSL` object in a separate process.
+
+Single process execution
+------------------------
+
+Let's start by demonstrating the issue with a single process where the main thread is
+hogging all of the CPU resources, leaving almost nothing for the
+:class:`~mne_lsl.player.PlayerLSL` thread to run and mock the LSL stream.
+
+.. GENERATED FROM PYTHON SOURCE LINES 20-35
+
+.. code-block:: Python
+
+
+
+    import multiprocessing as mp
+    import time
+
+    from mne.io import read_raw_fif
+
+    from mne_lsl.datasets import sample
+    from mne_lsl.player import PlayerLSL as Player
+
+    raw = read_raw_fif(sample.data_path() / "sample-ecg-raw.fif", preload=False)
+    raw.crop(0, 3).load_data()
+    player = Player(raw, chunk_size=200, n_repeat=1, name="example-process")
+    player
+
+
+
+
+
+.. rst-class:: sphx-glr-script-out
+
+ .. code-block:: none
+
+
+    <Player: example-process | OFF | /home/runner/mne_data/MNE-LSL-data/sample/sample-ecg-raw.fif>
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 37-50
+
+.. note::
+
+    Note that we use a ``chunk_size`` of 200 samples which does replicate a real-time
+    scenario where the LSL stream is usually using finer granularity (specific to the
+    device and LSL stream application). However, this large chunk size provides
+    stability to the documentation build on our CIs.
+
+The smaller the chunk size, the more resources are needed by the
+:class:`~mne_lsl.player.PlayerLSL` object to keep up with the LSL stream. This is why
+with a chunk size of 200 samples, we don't observe a difference between a main thread
+giving time to the player to run (scenario 1, while loop with :func:`time.sleep` calls
+to create idle time) and a main thread hogging all of the CPU resources (scenario 2,
+while loop without any idle time).
+
+.. GENERATED FROM PYTHON SOURCE LINES 50-70
+
+.. code-block:: Python
+
+
+    # scenario 1: give time to the player to run
+    start = time.time()
+    player.start()
+    while player.running:
+        time.sleep(0.5)
+    stop = time.time()
+    print(f"Elapsed time (free time): {stop - start:.2f} s")
+
+    # scenario 2: hog all of the CPU resources
+    start = time.time()
+    player.start()
[…]

[…]
