+++ .github/workflows/stubs.yaml
@@ -23,7 +23,7 @@ jobs:
       - run: uv pip install -e .[stubs] --quiet --system
       - run: mne-lsl sys-info --developer
       - run: python tools/stubgen.py
-      - uses: autofix-ci/action@v1.3.1
+      - uses: autofix-ci/action@v1.3.2
         if: ${{ github.event_name == 'pull_request' }}
       - name: Push stub files
         if: ${{ github.event_name == 'schedule' || github.event_name == 'workflow_dispatch' }}
