### 2025-06-13 9a93d8c
fix typo in docstring
