### 2025-06-14 5ba8fdc
use VS2019 toolset
