+++ 1.6.0/_downloads/d60a2acf61b276d4fc8fc253410a1fc4/10_peak_detection.py
@@ -1,0 +1,442 @@
+"""
+Real-time peak detection
+========================
+
+With a :class:`~mne_lsl.stream.StreamLSL`, we can build a real-time peak detector. The
+structure defined below will be adapted for cardiac R-peak detection, but remains
+valid for other peak detector. Note however that the peak detection performance, i.e.
+how fast it will be able to detect a peak, will depend heavily on the peak shape.
+
+For this example, consider a :class:`~mne_lsl.stream.StreamLSL` connected to an
+amplifier stream containing an ECG bipolar channel. We can detect in real-time the
+R-peak within the ECG signal. The objective of this example is to create a ``Detector``
+object able to detect new R-peak entering the buffer as fast as possible, with some
+robustness to external noise sources (e.g. movements) and a simple design.
+
+.. image:: ../../_static/tutorials/qrs.png
+    :align: center
+    :class: qrs-img
+
+First let's have a look to a sample ECG signal and to how we could detect the R-peak
+reliably with :func:`scipy.signal.find_peaks`.
+"""
+
+# sphinx_gallery_thumbnail_path = '_static/tutorials/qrs.png'
+
+import numpy as np
+from matplotlib import pyplot as plt
+from mne.io import read_raw_fif
+from mne.viz import set_browser_backend
+from scipy.signal import find_peaks
+
+from mne_lsl.datasets import sample
+
+raw = read_raw_fif(sample.data_path() / "sample-ecg-raw.fif", preload=True)
+raw
+
+# %%
+# This sample recording contains a single channel with the ECG signal.
+
+set_browser_backend("matplotlib")
+raw.plot(scalings=dict(misc=1300), show_scrollbars=False)
+plt.show()
+
+# %%
+# Filters
+# -------
+#
+# This recording is heavily affected by line noise (50 Hz in Europe). Our detector
+# should filter the signal to distinguish easily the QRS complex and the associated
+# R-peaks. Let's compare the raw signal with filtered signal using the following
+# settings:
+#
+# - notch at 50 and 100 Hz
+# - bandpass filter between 0.1 and 15 Hz
+# - lowpass filter at 15 Hz
+
+raw_notched = raw.copy()
+_ = raw_notched.notch_filter(50, method="iir", phase="forward", picks="misc")
+_ = raw_notched.notch_filter(100, method="iir", phase="forward", picks="misc")
+
+raw_bandpassed = raw.copy()
+_ = raw_bandpassed.filter(0.1, 15, method="iir", phase="forward", picks="misc")
+
+raw_lowpassed = raw.copy()
+_ = raw_lowpassed.filter(None, 15, method="iir", phase="forward", picks="misc")
+
+# %%
+# To compare those signals, it would be best if we could overlay them in a single plot.
+# Let's select a 5 seconds window and plot the detrended signals.
+#
+# .. note::
+#
+#     Contrary to the filter used by a :class:`~mne_lsl.stream.StreamLSL`, the forward
+#     filter in MNE don't use initial filter conditions. Thus, the beginning of the
+#     filter signal should not be used as the filter has not yet converged.
+
+start = int(120 * raw.info["sfreq"])
+stop = int(125 * raw.info["sfreq"])
+fig, ax = plt.subplots(1, 1, figsize=(10, 5), layout="constrained")
+for raw_, label in zip(
+    (raw, raw_notched, raw_bandpassed, raw_lowpassed),
+    ("raw", "notched", "bandpassed", "lowpassed"),
+):
+    data, times = raw_[:, start:stop]  # select 5 seconds
+    data -= data.mean()  # detrend
+    ax.plot(times, data.squeeze(), label=label)
+ax.legend()
+plt.show()
[…]
