+++ 1.3.1/_sources/changes/index.rst.txt
@@ -1,0 +1,10 @@
+Changelog
+=========
+
+.. toctree::
+    :titlesonly:
+
+    1.3.rst
+    1.2.rst
+    1.1.rst
+    1.0.rst

[…]
