### 2025-06-16 60e2203
Pin liblsl build to VS2019 toolset and update Windows install instructions (#440)
