+++ doc/development/changes/latest.rst
@@ -19,4 +19,6 @@
 Version 1.10
 ============
 
-- xxx
+- Add support for callbacks in :class:`~mne_lsl.stream.StreamLSL` objects with :meth:`~mne_lsl.stream.StreamLSL.add_callback` (:pr:`419` by `Mathieu Scheltienne`_)
+- Add support for hashing of :class:`~mne_lsl.stream.StreamLSL` objects (:pr:`424` by `Mathieu Scheltienne`_)
+- Build ``liblsl`` with the MSVC ``v142`` toolchain on Windows (:pr:`440` by `Mathieu Scheltienne`_)
