### 2025-06-16 8ebb3f1
Bump pypa/cibuildwheel from 2.23.3 to 3.0.0 (#441)
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
