### 2025-06-16 0ab5a9f
Update paper.bib
Following the example paper, https://joss.readthedocs.io/en/latest/example_paper.html, this grouping allows for the collective group name to appear and not be parsed as personal name
