+++ paper/paper.bib
@@ -36,7 +36,7 @@
 }
 
 @software{mne-realtime,
- author        = {MNE-realtime contributors},
+ author        = {{MNE-realtime contributors}},
  title         = {MNE-realtime},
  howpublished  = {\url{https://mne.tools/mne-realtime/}},
  note          = {Realtime analysis of MEG/EEG data with MNE. Accessed: 2024-08-08}
