+++ 1.1.1/_sources/index.rst.txt
@@ -1,0 +1,48 @@
+:hide-toc:
+
+.. include:: ./links.inc
+
+MNE-LSL
+=======
+
+Open-source Python package for real-time brain signal streaming framework
+based on the `Lab Streaming Layer (LSL) <lsl intro_>`_.
+
+Install
+-------
+
+``MNE-LSL`` is available on `PyPI <project pypi_>`_ and is distributed with a compatible
+version of `liblsl <lsl lib c++_>`_.
+
+.. tab-set::
+
+    .. tab-item:: PyPI
+
+        .. code-block:: console
+
+            $ pip install mne-lsl
+
+    .. tab-item:: Source
+
+        .. code-block:: console
+
+            $ pip install git+https://github.com/mne-tools/mne-lsl
+
+Supporting institutions
+-----------------------
+
+The development of ``MNE-LSL`` is supported by the
+`Human Neuroscience Platform, Fondation Campus Biotech Geneva <fcbg hnp_>`_.
+
+.. image:: _static/partners/fcbg-hnp-meeg.png
+    :alt: FCBG - HNP - MEEG/BCI Platform
+    :width: 150
+
+.. toctree::
+    :hidden:
+
+    resources/install.rst
+    api/index.rst
+    resources/command_line.rst
+    generated/tutorials/index.rst
+    changes/index.rst
