+++ 1.7.0/_sources/generated/examples/40_decode.rst.txt
@@ -1,0 +1,1546 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "generated/examples/40_decode.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_generated_examples_40_decode.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_generated_examples_40_decode.py:
+
+
+Decoding real-time data
+=======================
+
+.. include:: ./../../links.inc
+
+This example demonstrates how to decode real-time data using `MNE-Python <mne stable_>`_
+and `Scikit-learn <sklearn stable_>`_. We will stream the ``sample_audvis_raw.fif``
+file from MNE's sample dataset with a :class:`~mne_lsl.player.PlayerLSL`, process the
+signal through a :class:`~mne_lsl.stream.StreamLSL`, and decode the epochs created with
+:class:`~mne_lsl.stream.EpochsStream`.
+
+.. GENERATED FROM PYTHON SOURCE LINES 13-37
+
+.. code-block:: Python
+
+
+    import time
+    import uuid
+
+    import numpy as np
+    from matplotlib import pyplot as plt
+    from mne.decoding import Vectorizer
+    from mne.io import read_raw_fif
+    from sklearn.linear_model import LogisticRegression
+    from sklearn.model_selection import ShuffleSplit, cross_val_score
+    from sklearn.pipeline import Pipeline
+    from sklearn.preprocessing import StandardScaler
+
+    from mne_lsl.datasets import sample
+    from mne_lsl.player import PlayerLSL
+    from mne_lsl.stream import EpochsStream, StreamLSL
+
+    fname = sample.data_path() / "mne-sample" / "sample_audvis_raw.fif"
+    raw = read_raw_fif(fname, preload=False).pick(("meg", "stim")).load_data()
+    source_id = uuid.uuid4().hex
+    player = PlayerLSL(raw, chunk_size=200, name="real-time-decoding", source_id=source_id)
+    player.start()
+    player.info
+
+
+
+
+
+
+.. raw:: html
+
+    <div class="output_subarea output_html rendered_html output_result">
+    <script type="text/javascript">
+        const toggleVisibility = (className) => {
+
+      const elements = document.querySelectorAll(`.${className}`)
+
+      elements.forEach(element => {
+        if (element.classList.contains('repr-section-header')) {
+          // Don't collapse the section header row.
+           return
+        }
+        if (element.classList.contains('repr-element-collapsed')) {
+          // Force a reflow to ensure the display change takes effect before removing the class
+          element.classList.remove('repr-element-collapsed')
+          element.offsetHeight // This forces the browser to recalculate layout
[…]
