### 2025-06-16 #8088 @teonbrooks
General comments:

I've now had the chance to work my way through this software package and paper and I am very satisfied with its state. It's well-documented, the examples are clear and user-friendly. The code provides a comprehensive test suite.

The mock client is very useful to understand what's going on with the real-time streaming even if you're not using a device. I had a chance to connect my Neurosity EEG headset I own and was able to use the package to stream EEG data into my terminal.

I've submitted some suggestions and corrections to the package so my concerns have been addressed.

The package is a great contribution to the community.
