### 2025-06-22 #8088 @matthiasdold
I too consider the project in a very nice shape, both in terms of offered functionality and documentation.

All examples and ... worked on my machine (MacOS 15.5, M1 Pro, ).

I worked through all examples and streamed data of various file sources including. The project has also proven very valuable for teaching purposes. It allows to create very simple examples of `mock-up` closed-loop processing that can be explained to students due to the lean code required by the module.
