### 2025-06-23 mne-tools/mne-lsl#426 @matthiasdold
Perfect, thanks @mscheltienne, I will check it out ASAP.
