### 2025-06-23 #8088 @teonbrooks
> The paper is concise but covers the key contributions of mne and LSL with according references. While the discussion of other platforms is rather short, not mentioning e.g., [BCI2000](/www.bci2000.org), I consider this justifiable as the core functionality claim is to make the MNE python tooling available to closed-loop processing via LSL streaming data.

that's a good point, it might be worth to mention in the summary overview as one of the other options in the field but you can emphasize that the focus is on the use of LSL based on the reasons you provide
