### 2025-06-24 8ff8e36
[pre-commit.ci] pre-commit autoupdate (#445)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@dandelion.science>
