+++ 1.7.1/_downloads/2d736980625d98437c359cae45dc7a95/40_epochs.py
@@ -1,0 +1,295 @@
+"""
+Epoching a Stream in real-time
+==============================
+
+.. include:: ./../../links.inc
+
+The :class:`~mne_lsl.stream.EpochsStream` object can be used similarly to
+:class:`mne.Epochs` to create epochs from a continuous stream of samples around events
+of interest.
+
+.. note::
+
+    The :class:`~mne_lsl.stream.EpochsStream` object is designed to work with
+    any ``Stream`` object. At the time of writing, only
+    :class:`~mne_lsl.stream.StreamLSL` is available, but any object inheriting from the
+    abstract :class:`~mne_lsl.stream.BaseStream` object should work.
+
+A :class:`~mne_lsl.stream.EpochsStream` object support peak-to-peak rejection, baseline
+correction and detrending.
+
+Event source
+------------
+
+A :class:`~mne_lsl.stream.EpochsStream` object requires an event source to create
+epochs. 3 event sources are supported:
+
+- a set of ``'stim'`` channels within the attached ``Stream`` object.
+- a set of ``'stim'`` channels within a separate ``Stream`` object.
+- an irregularly sampled ``Stream`` object.
+
+.. note::
+
+    In the case of an irregularly sampled ``Stream`` object, only numerical streams are
+    supported at the moment because interaction with ``str`` streams in Python is not
+    as performant as interaction with numerical streams.
+
+Set of ``'stim'`` channels
+--------------------------
+
+The set of ``'stim'`` channels from which the events are extracted can be either part
+of the regularly sampled ``Stream`` object epoched (argument ``stream``) or part of a
+separate regularly sampled ``Stream`` object (argument ``event_stream``). In both case,
+the channel(s) type should be ``'stim'`` and the channel(s) should be formatted for
+:func:`mne.find_events` to correctly extract the events. The channels to consider are
+provided in the argument ``event_channels`` and the events to consider in the argument
+``event_id``. Let's create epochs around the event ID ``2`` from the ``'STI 014'``
+channel of MNE's sample dataset.
+"""
+
+import time
+import uuid
+
+import numpy as np
+from matplotlib import pyplot as plt
+from mne import Epochs, EpochsArray, annotations_from_events, find_events
+from mne.io import read_raw_fif
+
+from mne_lsl.datasets import sample
+from mne_lsl.lsl import resolve_streams
+from mne_lsl.player import PlayerLSL
+from mne_lsl.stream import EpochsStream, StreamLSL
+
+fname = sample.data_path() / "mne-sample" / "sample_audvis_raw.fif"
+raw = read_raw_fif(fname, preload=False).pick(("meg", "stim")).load_data()
+source_id = uuid.uuid4().hex
+player = PlayerLSL(
+    raw,
+    chunk_size=200,
+    name="tutorial-epochs-1",
+    source_id=source_id,
+    annotations=False,
+).start()
+player.info
+
+# %%
+# .. note::
+#
+#     A ``chunk_size`` of 200 samples is used here to ensure stability and reliability
+#     while building the documentation on the CI. In practice, a ``chunk_size`` of 200
+#     samples is too large to represent a real-time application.
+#
+# In the cell above, a mock LSL stream is created using the ``'meg'`` and ``'stim'``
+# channels of MNE's sample dataset. Now, we need to create a
+# :class:`~mne_lsl.stream.StreamLSL` object connected to this mock LSL stream. The
+# channel ``"MEG 2443"`` is marked as bad and the signal is filtered with a low-pass
+# filter.
+
[…]
