+++ 1.5.0/_sources/generated/examples/sg_execution_times.rst.txt
@@ -1,0 +1,49 @@
+
+:orphan:
+
+.. _sphx_glr_generated_examples_sg_execution_times:
+
+
+Computation times
+=================
+**03:56.211** total execution time for 5 files **from generated/examples**:
+
+.. container::
+
+  .. raw:: html
+
+    <style scoped>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
+    </style>
+    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
+    <script type="text/javascript" class="init">
+    $(document).ready( function () {
+        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
+    } );
+    </script>
+
+  .. list-table::
+   :header-rows: 1
+   :class: table table-striped sg-datatable
+
+   * - Example
+     - Time
+     - Mem (MB)
+   * - :ref:`sphx_glr_generated_examples_40_decode.py` (``40_decode.py``)
+     - 01:19.426
+     - 626.9
+   * - :ref:`sphx_glr_generated_examples_30_real_time_evoked_responses.py` (``30_real_time_evoked_responses.py``)
+     - 00:59.956
+     - 623.6
+   * - :ref:`sphx_glr_generated_examples_20_bandpower.py` (``20_bandpower.py``)
+     - 00:48.897
+     - 235.9
+   * - :ref:`sphx_glr_generated_examples_10_peak_detection.py` (``10_peak_detection.py``)
+     - 00:38.199
+     - 192.8
+   * - :ref:`sphx_glr_generated_examples_00_player_separate_process.py` (``00_player_separate_process.py``)
+     - 00:09.733
+     - 188.8

[…]
