### 2025-06-24 mne-tools/mne-lsl#426 @matthiasdold
I could reproduce the figures with the concise code you provided  - nice and easy to reproduce 👍
