### 2025-06-27 6a7dc8f
add warning and support for too small epoch bufsize
