### 2025-06-27 44fca6a
Fix bug where epoch with too small buffer could fail due to the event shape (#447)
