+++ 1.4.0/_sources/generated/examples/index.rst.txt
@@ -1,0 +1,101 @@
+:orphan:
+
+.. include:: ../../links.inc
+
+Examples
+========
+
+This page contains examples using ``MNE-LSL``.
+
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbnails">
+
+.. thumbnail-parent-div-open
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="With a PlayerLSL object, you can replay an MNE-Python readable file as a valid LSL stream. This...">
+
+.. only:: html
+
+  .. image:: /generated/examples/images/thumb/sphx_glr_00_player_separate_process_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_generated_examples_00_player_separate_process.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Replaying a file in a separate process</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="With a StreamLSL, we can build a real-time peak detector. The structure defined below will be a...">
+
+.. only:: html
+
+  .. image:: /generated/examples/images/thumb/sphx_glr_10_peak_detection_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_generated_examples_10_peak_detection.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Real-time peak detection</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="With a StreamLSL, we can compute the bandpower on a time rolling window. For this example, we w...">
+
+.. only:: html
+
+  .. image:: /generated/examples/images/thumb/sphx_glr_20_bandpower_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_generated_examples_20_bandpower.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Bandpower rolling window</div>
+    </div>
+
+
+.. thumbnail-parent-div-close
+
+.. raw:: html
+
+    </div>
+
+
+.. toctree::
+   :hidden:
+
+   /generated/examples/00_player_separate_process
+   /generated/examples/10_peak_detection
+   /generated/examples/20_bandpower
+
+
+.. only:: html
+
+  .. container:: sphx-glr-footer sphx-glr-footer-gallery
+
+    .. container:: sphx-glr-download sphx-glr-download-python
+
+      :download:`Download all examples in Python source code: examples_python.zip </generated/examples/examples_python.zip>`
+
+    .. container:: sphx-glr-download sphx-glr-download-jupyter
+
+      :download:`Download all examples in Jupyter notebooks: examples_jupyter.zip </generated/examples/examples_jupyter.zip>`
+
+
+.. only:: html
+
+ .. rst-class:: sphx-glr-signature
+
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
