+++ 1.0.0/_sources/generated/tutorials/sg_execution_times.rst.txt
@@ -1,0 +1,19 @@
+
+:orphan:
+
+.. _sphx_glr_generated_tutorials_sg_execution_times:
+
+
+Computation times
+=================
+**00:35.631** total execution time for **generated_tutorials** files:
+
++-----------------------------------------------------------------------------------------+-----------+----------+
+| :ref:`sphx_glr_generated_tutorials_10_player.py` (``10_player.py``)                     | 00:13.127 | 253.1 MB |
++-----------------------------------------------------------------------------------------+-----------+----------+
+| :ref:`sphx_glr_generated_tutorials_00_stream.py` (``00_stream.py``)                     | 00:11.626 | 120.2 MB |
++-----------------------------------------------------------------------------------------+-----------+----------+
+| :ref:`sphx_glr_generated_tutorials_30_timestamps.py` (``30_timestamps.py``)             | 00:06.306 | 101.8 MB |
++-----------------------------------------------------------------------------------------+-----------+----------+
+| :ref:`sphx_glr_generated_tutorials_20_stream_meas_info.py` (``20_stream_meas_info.py``) | 00:04.572 | 14.5 MB  |
++-----------------------------------------------------------------------------------------+-----------+----------+
