### 2025-07-11 #8088 @mscheltienne
@teonbrooks @matthiasdold Friendly ping :) Would you like https://github.com/mne-tools/mne-lsl/pull/449 merged as is, slightly changed, ignored and closed? It seems that this might be the last blocker we have!
