### 2025-07-11 #8088 @matthiasdold
Thanks a lot for the ping @mscheltienne - https://github.com/mne-tools/mne-lsl/pull/449 looks good to me and indeed would then capture everything from my side.
