### 2025-07-11 e543c13
Mention BCI2000 (#449)
