+++ 1.2.0/_downloads/3398261e6359dbd9878364b1fc3b1009/10_low_level_API.py
@@ -1,0 +1,171 @@
+"""
+.. _tut-low-level-api:
+
+Low-level LSL API
+=================
+
+.. include:: ./../../links.inc
+
+.. _lsl language bindings: https://github.com/sccn/labstreaminglayer/tree/master/LSL
+
+LSL is a library designed for streaming time series data across different platforms and
+programming languages. The `core library <lsl lib_>`_ is primarily written in C++, and
+bindings are accessible for Python, C#, Java, MATLAB, and Unity, among others. You can
+find a comprehensive list `here <lsl language bindings_>`_.
+
+MNE-LSL provides a reimplementation of the `python binding <lsl python_>`_, known as
+``pylsl``, within the ``mne_lsl.lsl`` module. It introduces additional functionalities
+to simplify the low-level interaction with LSL streams. Moreover, it enhances the
+detection of liblsl on your system and can retrieve a compatible version online if
+necessary. The differences between ``pylsl`` and ``mne_lsl.lsl`` are detailed
+:ref:`here<resources/pylsl:Differences with pylsl>`.
+"""
+
+# sphinx_gallery_thumbnail_path = '_static/tutorials/lsl.png'
+
+import time
+import uuid
+
+import numpy as np
+
+from mne_lsl.lsl import (
+    StreamInfo, StreamInlet, StreamOutlet, local_clock, resolve_streams
+)
+
+# %%
+# Creating a stream
+# -----------------
+#
+# To create a stream, you must first define its properties. This is achieved by creating
+# a :class:`~mne_lsl.lsl.StreamInfo` object, which specifies the stream's name, type,
+# source and properties. Convenience methods are available to set the channel
+# properties, including :meth:`~mne_lsl.lsl.StreamInfo.set_channel_info`, which uses a
+# :class:`mne.Info` object as source.
+
+sinfo = StreamInfo(
+    name="my-stream",
+    stype="eeg",
+    n_channels=3,
+    sfreq=1024,
+    dtype="float32",
+    source_id=uuid.uuid4().hex[:6],
+)
+sinfo.set_channel_names(["Fz", "Cz", "Oz"])
+sinfo.set_channel_types("eeg")
+sinfo.set_channel_units("microvolts")
+
+# %%
+# Once the :class:`~mne_lsl.lsl.StreamInfo` object is created, a
+# :class:`~mne_lsl.lsl.StreamOutlet` can be instantiated to create the stream.
+
+outlet = StreamOutlet(sinfo)
+
+# %%
+# Discover streams
+# ----------------
+#
+# At this point, the :class:`~mne_lsl.lsl.StreamOutlet` is available on the network. The
+# function :func:`~mne_lsl.lsl.resolve_streams` discovers all available streams on the
+# network. This operation is commonly named the stream resolution.
+#
+# .. note::
+#
+#     The stream resolution can be restricted by providing the ``name``, ``stype``, and
+#     ``source_id`` arguments.
+
+streams = resolve_streams()
+assert len(streams) == 1
+streams[0]
+
+# %%
+# The resolution retrieves only the stream basic properties. The channel properties,
+# stored in the stream description in an XML element tree, are absent from a
+# :class:`~mne_lsl.lsl.StreamInfo` returned by the resolution function.
+
+assert streams[0].get_channel_names() is None
+
+# %%
+# Connect to a Stream
+# -------------------
+#
+# To connect to a stream, a :class:`~mne_lsl.lsl.StreamInlet` object must be created
+# using the resolved :class:`~mne_lsl.lsl.StreamInfo`. Once the stream is opened with
[…]
