+++ 1.2.0/_sources/api/utilities.rst.txt
@@ -1,0 +1,21 @@
+Utilities
+---------
+
+.. currentmodule:: mne_lsl
+
+Logging utilities are available to interact with ``mne_lsl`` logger.
+
+.. autosummary::
+   :toctree: ../generated/api
+   :nosignatures:
+
+    add_file_handler
+    set_log_level
+
+Development utilities are available to help debug a setup.
+
+.. autosummary::
+   :toctree: ../generated/api
+   :nosignatures:
+
+    sys_info
