### 2025-07-14 779b900
add test and docstring for new feature
