### 2025-07-15 c0ea80f
add initial support for event id on irregularly sampled stream
