### 2025-07-15 8ff09c8
[pre-commit.ci] pre-commit autoupdate (#453)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@dandelion.science>
