+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
     - repo: https://github.com/astral-sh/ruff-pre-commit
-      rev: v0.12.1
+      rev: v0.12.3
       hooks:
           - id: ruff
             name: ruff linter
