### 2025-07-15 fb4b023
Add support for `event_id` in irregularly sampled event stream (#455)
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
