+++ tests/stream/test_epochs.py
@@ -10,7 +10,7 @@ import pytest
 from mne import annotations_from_events, create_info, find_events
 from mne.io import RawArray, read_raw_fif
 from mne.io.base import BaseRaw
-from numpy.testing import assert_allclose
+from numpy.testing import assert_allclose, assert_array_equal
 
 from mne_lsl.datasets import testing
 from mne_lsl.lsl import StreamInfo, StreamOutlet
@@ -788,8 +788,7 @@ def test_ensure_event_id_with_event_stream(
         10, stype="annotations", source_id=mock_lsl_stream_with_annotations.source_id
     ).connect(acquisition_delay=0.1)
     assert _ensure_event_id(None, event_stream) is None
-    with pytest.warns(RuntimeWarning, match="should be set to None"):
-        _ensure_event_id(dict(event=1), event_stream)
+    assert _ensure_event_id(dict(event=1), event_stream) == dict(event=1)
     event_stream.disconnect()
 
 
@@ -1133,3 +1132,122 @@ def test_epochs_with_more_events_than_buffer_size(
         while epochs.n_new_epochs != 5 and time.monotonic() - start < 3:
             epochs.acquire()
             time.sleep(0.5)
+
+
+def test_epochs_with_irregular_numerical_event_stream_and_event_id(
+    mock_lsl_stream: DummyPlayer, outlet_marker: StreamOutlet
+) -> None:
+    """Test the selection of epochs based on event_id."""
+    stream = StreamLSL(
+        10, name=mock_lsl_stream.name, source_id=mock_lsl_stream.source_id
+    ).connect(acquisition_delay=0.1)
+    sinfo = outlet_marker.get_sinfo()
+    event_stream = StreamLSL(5, name=sinfo.name, source_id=sinfo.source_id).connect(
+        acquisition_delay=0.1
+    )
+    epochs = EpochsStream(
+        stream,
+        bufsize=10,
+        event_id=1,
+        event_channels="marker",
+        event_stream=event_stream,
+        tmin=-0.5,
+        tmax=0,
+        baseline=None,
+        picks="eeg",
+    ).connect(acquisition_delay=None)
+    assert epochs.n_new_epochs == 0
+    epochs.acquire()
+    assert epochs.n_new_epochs == 0
+    time.sleep(0.5)
+    # push 5 events, only 3 with ID 1
+    outlet_marker.push_sample(np.array([1], dtype=sinfo.dtype))
+    time.sleep(0.1)
+    outlet_marker.push_sample(np.array([2], dtype=sinfo.dtype))
+    time.sleep(0.1)
+    outlet_marker.push_sample(np.array([1], dtype=sinfo.dtype))
+    time.sleep(0.1)
+    outlet_marker.push_sample(np.array([3], dtype=sinfo.dtype))
+    time.sleep(0.1)
+    outlet_marker.push_sample(np.array([1], dtype=sinfo.dtype))
+    time.sleep(0.1)
+    start = time.monotonic()
+    while epochs.n_new_epochs != 3 and time.monotonic() - start < 3:
+        epochs.acquire()
+        time.sleep(0.5)
+    # check that we got 3 epochs on the code '1'
+    events = epochs.events
+    assert_array_equal(events[events != 0], [1, 1, 1])
+
+
+@pytest.fixture
+def outlet_marker_3_channel() -> Generator[StreamOutlet, None, None]:
+    """Stream outlet for a marker stream with 2 channels."""
+    sinfo = StreamInfo(
+        name="events",
+        stype="marker",
+        n_channels=3,
+        sfreq=0,
+        dtype=np.int8,
[…]
