### 2025-07-15 32d0f57
handle post-release 1.10
