### 2025-07-15 ac83968
update documentation for release 1.10
