+++ stable/_sources/generated/tutorials/00_introduction.rst.txt
@@ -297,9 +297,9 @@ of the file is reached.
 
 
 
-    <tr class="mne-repr-section-header general-2fc7376b-944f-4e19-b856-14bc97cfcadc"
+    <tr class="mne-repr-section-header general-8368e9c1-4430-44ce-a8aa-d18af207525c"
          title="Hide section" 
-        onclick="toggleVisibility('general-2fc7376b-944f-4e19-b856-14bc97cfcadc')">
+        onclick="toggleVisibility('general-8368e9c1-4430-44ce-a8aa-d18af207525c')">
         <th class="mne-repr-section-toggle">
             <button >
                 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><!--!Font Awesome Free 6.6.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path d="M233.4 406.6c12.5 12.5 32.8 12.5 45.3 0l192-192c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L256 338.7 86.6 169.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l192 192z"/></svg>
@@ -311,26 +311,26 @@ of the file is reached.
     </tr>
 
 
-    <tr class="repr-element general-2fc7376b-944f-4e19-b856-14bc97cfcadc ">
+    <tr class="repr-element general-8368e9c1-4430-44ce-a8aa-d18af207525c ">
         <td class="mne-repr-section-toggle"></td>
         <td>MNE object type</td>
         <td>Info</td>
     </tr>
-    <tr class="repr-element general-2fc7376b-944f-4e19-b856-14bc97cfcadc ">
+    <tr class="repr-element general-8368e9c1-4430-44ce-a8aa-d18af207525c ">
         <td class="mne-repr-section-toggle"></td>
         <td>Measurement date</td>
     
         <td>Unknown</td>
     
     </tr>
-    <tr class="repr-element general-2fc7376b-944f-4e19-b856-14bc97cfcadc ">
+    <tr class="repr-element general-8368e9c1-4430-44ce-a8aa-d18af207525c ">
         <td class="mne-repr-section-toggle"></td>
         <td>Participant</td>
     
         <td>Unknown</td>
     
     </tr>
-    <tr class="repr-element general-2fc7376b-944f-4e19-b856-14bc97cfcadc ">
+    <tr class="repr-element general-8368e9c1-4430-44ce-a8aa-d18af207525c ">
         <td class="mne-repr-section-toggle"></td>
         <td>Experimenter</td>
     
@@ -345,9 +345,9 @@ of the file is reached.
 
 
 
-    <tr class="mne-repr-section-header acquisition-c35b208f-7dce-4671-9afc-fa4509cab7c8"
+    <tr class="mne-repr-section-header acquisition-4e638f55-cbca-491b-9117-805df57d5526"
          title="Hide section" 
-        onclick="toggleVisibility('acquisition-c35b208f-7dce-4671-9afc-fa4509cab7c8')">
+        onclick="toggleVisibility('acquisition-4e638f55-cbca-491b-9117-805df57d5526')">
         <th class="mne-repr-section-toggle">
             <button >
[…]
