+++ stable/_sources/generated/tutorials/40_epochs.rst.txt
@@ -250,9 +250,9 @@ channel of MNE's sample dataset.
 
 
 
-    <tr class="mne-repr-section-header general-44cdeb99-b5a4-4feb-a481-a87099f58f19"
+    <tr class="mne-repr-section-header general-9ad46e98-4770-4638-a79f-15644936236d"
          title="Hide section" 
-        onclick="toggleVisibility('general-44cdeb99-b5a4-4feb-a481-a87099f58f19')">
+        onclick="toggleVisibility('general-9ad46e98-4770-4638-a79f-15644936236d')">
         <th class="mne-repr-section-toggle">
             <button >
                 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><!--!Font Awesome Free 6.6.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path d="M233.4 406.6c12.5 12.5 32.8 12.5 45.3 0l192-192c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L256 338.7 86.6 169.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3l192 192z"/></svg>
@@ -264,26 +264,26 @@ channel of MNE's sample dataset.
     </tr>
 
 
-    <tr class="repr-element general-44cdeb99-b5a4-4feb-a481-a87099f58f19 ">
+    <tr class="repr-element general-9ad46e98-4770-4638-a79f-15644936236d ">
         <td class="mne-repr-section-toggle"></td>
         <td>MNE object type</td>
         <td>Info</td>
     </tr>
-    <tr class="repr-element general-44cdeb99-b5a4-4feb-a481-a87099f58f19 ">
+    <tr class="repr-element general-9ad46e98-4770-4638-a79f-15644936236d ">
         <td class="mne-repr-section-toggle"></td>
         <td>Measurement date</td>
     
         <td>2002-12-03 at 19:01:10 UTC</td>
     
     </tr>
-    <tr class="repr-element general-44cdeb99-b5a4-4feb-a481-a87099f58f19 ">
+    <tr class="repr-element general-9ad46e98-4770-4638-a79f-15644936236d ">
         <td class="mne-repr-section-toggle"></td>
         <td>Participant</td>
     
         <td>Unknown</td>
     
     </tr>
-    <tr class="repr-element general-44cdeb99-b5a4-4feb-a481-a87099f58f19 ">
+    <tr class="repr-element general-9ad46e98-4770-4638-a79f-15644936236d ">
         <td class="mne-repr-section-toggle"></td>
         <td>Experimenter</td>
     
@@ -298,9 +298,9 @@ channel of MNE's sample dataset.
 
 
 
-    <tr class="mne-repr-section-header acquisition-b1713ac1-e546-4556-b5d5-8d8d00a6823d"
+    <tr class="mne-repr-section-header acquisition-02e9aeb7-cd9e-4b0c-abaa-a356ff74eed7"
          title="Hide section" 
-        onclick="toggleVisibility('acquisition-b1713ac1-e546-4556-b5d5-8d8d00a6823d')">
+        onclick="toggleVisibility('acquisition-02e9aeb7-cd9e-4b0c-abaa-a356ff74eed7')">
         <th class="mne-repr-section-toggle">
             <button >
[…]
