+++ stable/_sources/development/changes/index.rst.txt
@@ -4,6 +4,7 @@ Changelog
 .. toctree::
     :titlesonly:
 
+    1.10.rst
     1.9.rst
     1.8.rst
     1.7.rst
