+++ stable/_sources/development/changes/1.9.rst.txt
@@ -14,8 +14,6 @@
 
 .. include:: ./authors.inc
 
-.. _latest:
-
 Version 1.9
 ===========
 

[…]
