### 2025-07-21 90bad43
Minor editing fixes for JOSS paper (#456)
