+++ paper/paper.bib
@@ -10,8 +10,8 @@
 }
 
 @misc{brainflow,
-  author       = {BrainFlow},
-  title        = {BrainFlow is a library intended to obtain, parse and analyze EEG, EMG, ECG and other kinds of data from biosensors.},
+  author       = {BrainFlow contributors},
+  title        = {BrainFlow},
   howpublished = {\url{https://brainflow.org}},
   note         = {Accessed: 2024-08-08}
 }
@@ -39,10 +39,12 @@
  year          = {2020}
 }
 
-@software{LSL,
-  author       = {Stenner, Tristan and Boulay, Chadwick and others},
+@misc{LSL,
+  author       = {{LabStreamingLayer contributors}},
   title        = {LabStreamingLayer},
-  howpublished = {\url{https://github.com/sccn/labstreaminglayer}},
+  publisher    = {GitHub},
+  journal      = {GitHub repository},
+  url          = {https://github.com/sccn/labstreaminglayer},
   note         = {LSL is an open-source networked middleware ecosystem to stream, receive, synchronize, and record neural, physiological, and behavioral data streams acquired from diverse sensor hardware. Accessed: 2025-05-28}
 }
 
@@ -62,15 +64,17 @@
 }
 
 @software{openbci,
- author        = {OpenBCI},
+ author        = {{OpenBCI contributors}},
  title         = {OpenBCI},
  howpublished  = {\url{https://openbci.com}},
  note          = {OpenBCI creates open-source tools for biosensing and neuroscience. Accessed: 2024-08-08}
 }
 
-@software{pylsl,
-  author       = {Boulay, Chadwick and Stenner, Tristan and others},
+@misc{pylsl,
+  author       = {{pyLSL contributors}},
   title        = {pyLSL},
-  howpublished = {\url{https://github.com/labstreaminglayer/pylsl}},
+  publisher    = {GitHub},
+  journal      = {GitHub repository},
+  url          = {https://github.com/labstreaminglayer/pylsl},
   note         = {Python interface to the Lab Streaming Layer. Accessed: 2024-08-08}
 }
