+++ doc/development/changes/1.10.rst
@@ -16,8 +16,13 @@
 
 .. _latest:
 
-Version 1.10
-============
+Version 1.10.1
+==============
+
+- Release for JOSS publication and associated DOI
+
+Version 1.10.0
+==============
 
 - Add support for callbacks in :class:`~mne_lsl.stream.StreamLSL` objects with :meth:`~mne_lsl.stream.StreamLSL.add_callback` (:pr:`419` by `Mathieu Scheltienne`_)
 - Add support for hashing of :class:`~mne_lsl.stream.StreamLSL` objects (:pr:`424` by `Mathieu Scheltienne`_)
