### 2025-07-22 #8088 @mstimberg
@mscheltienne thanks for the update, this all looks good so far. Could you please change the title of the Zenodo archive so that it matches the title of the paper? I know that it is a bit annoying and does not fit with the automatically generated archives, but it is the preferred JOSS style for know. (Just to be clear: this does not need a new release, you can change the title of the existing release on Zenodo).
