### 2025-07-22 #8088 @mstimberg
Great, thanks. Very minor nitpicking: there is a small mismatch in the affiliations between the paper and the Zenodo archive. On Zenodo, Eric Larson's affiliation is Institute**s** (plural), and the affiliation for Kyuhwa Lee is missing the "Geneva, Switzerland" part.
