### 2025-07-23 4d9a363
Fix Brainflow contributors in papers.bib (#458)
