+++ paper/paper.bib
@@ -10,7 +10,7 @@
 }
 
 @misc{brainflow,
-  author       = {BrainFlow contributors},
+  author       = {{BrainFlow contributors}},
   title        = {BrainFlow},
   howpublished = {\url{https://brainflow.org}},
   note         = {Accessed: 2024-08-08}
