### 2025-07-23 #8088 @mscheltienne
No problem, I'll cut 1.10.2 and will report back, the release process will take probably ~1h.
