### 2025-07-23 #8088 @mstimberg
JOSS policy is that if changes only concern the paper, it is ok if they are not reflected in the release, since the paper is archived separately.
